% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_mod.R
\name{adj.mod}
\alias{adj.mod}
\title{Compute Adjacency Matrix and Identify Modules}
\usage{
adj.mod(data, type, minSize = 15)
}
\arguments{
\item{data}{The processed data matrix resulting from the function "filter.data()".}

\item{type}{There are two options in this argument, "signed" and "unsigned". "signed" means both positive and negative adjacency between genes are maintained in network module identification while "unsigned" takes the absolute values of negative adjacency. Refer to "WGCNA" for more details.}

\item{minSize}{The minimum module size in module identification. The default value is 15. Refer to "WGCNA" for more details.}
}
\value{
The adjacency matrix file "adj.txt" and module assignment file "mod.txt", which are all automatically saved in the directory "local_mode_result."
}
\description{
It designed to compute adjacency matrix and define modules. The resulting adjacency matrix (adj.txt) and module definition (mod.txt) files are saved in the directory "local_mode_result", both of which should be uploaded to "Compute locally" in the Shiny app, which is launched by running "spatial.hm.all()".
}
\section{Details}{

This function relies on "WGCNA". First, it computes the adjacency matrix based on the input data matrix, then converts the resulting adjacency matrix to a scale-free topology using a soft threshold (sft). If the type is "signed" sft=12, if the type is "unsigned" sft=6. Next, the modules are identified using dynamic tree cutting algorithm at four sensitivity levels 0, 1, 2, 3. From 0 to 3, more modules are identified but module sizes are decreasing. In order to enhance the performance of this Spatial Heatmap, modules are only identified at level 2 and 3, since the interactive network functionality of this app works better on smaller modules. Refer to "WGCNA" for more details.
}

\examples{
 
data.path <- system.file("extdata/example", "gene_expr_ann_row_gen.txt", package = "spatialHeatmap")
exp <- filter.data(data=data.path, sep="\\t", isRowGen=TRUE, c(0, 0), c(0.1, 10000), "processed_data")
adj_mod <- adj.mod(data=exp, type="signed", minSize=15)
}
\references{
Langfelder P and Horvath S, WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics 2008, 9:559 doi:10.1186/1471-2105-9-559 \cr Peter Langfelder, Steve Horvath (2012). Fast R Functions for Robust Correlations and Hierarchical Clustering. Journal of Statistical Software, 46(11), 1-17. URL http://www.jstatsoft.org/v46/i11/ \cr Peter Langfelder, Steve Horvath (2012). Fast R Functions for Robust Correlations and Hierarchical Clustering. Journal of Statistical Software, 46(11), 1-17. URL http://www.jstatsoft.org/v46/i11/.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
