% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_dataset.R
\name{auc_dataset}
\alias{auc_dataset}
\title{Extracting AUCs by data sets}
\usage{
auc_dataset(wk.dir, min.true = 300, min.total = 500, min.auc = 0.5)
}
\arguments{
\item{wk.dir}{The directory containing co-clustering results that are generated by \code{coclus_opt}.}
}
\value{
An object of ggplot.
}
\description{
After coclustering optimization, visualize AUCs generated by optimal parameter settings on validating data sets.
}
\examples{
# Create parameter settings and corresponding aucs.
df.auc <- data.frame(auc=runif(n=5, min=1e-12, max=.99))
ran <- seq(0.2, 0.8, 0.1)
df.spd <- data.frame(sim=sample(ran, 5, replace=FALSE), sim.p=sample(ran, 5, replace=FALSE), dim=sample(seq(5, 40, 1), 5, replace=FALSE))
df.spd$spd.set <- paste0('s', df.spd$sim, 'p', df.spd$sim.p, 'd', df.spd$dim)

df.auc <- cbind(df.spd, df.auc)
# Plot the aucs.
auc_bar(df.auc, auc='auc', thr=0.5) 
# See package vignette by calling "browseVignettes('spatialHeatmap')".
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
