% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_stat.R
\name{auc_stat}
\alias{auc_stat}
\title{Extract AUC statistics in coclustering optimization}
\usage{
auc_stat(
  wk.dir,
  tar.par = "norm",
  total.min = 500,
  true.min = 300,
  aucs = round(seq(0.5, 0.9, 0.1), 1)
)
}
\arguments{
\item{wk.dir}{The working directory of coclustering.}

\item{tar.par}{The target parameter in optimization, one of \code{norm}, \code{filter}, \code{dimred}, \code{graph}, and \code{spd.set} corresponding to normalization methods, filtering parameter sets, dimensionality reduction methods, graph-building methods, and \code{sim}/\code{sim.p}/\code{dim} sets respectively.}

\item{total.min, true.min, aucs}{Cutoffs to extract AUCs. AUCs over a cutoff and having total bulk tissue assignments above \code{total} and true assignments above \code{true.min} are extracted. The default is \code{total.min=500}, \code{true.min=300}, \code{round(seq(0.5, 0.9, 0.1), 1)}. Each AUC cuttoff in \code{aucs} is used independently.}
}
\value{
An nested \code{list}.
}
\description{
Extract AUC statistics of target parameters in coclustering optimization.
}
\examples{
See function "coclus_opt" by calling "?coclus_opt".
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
