% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_violin.R
\name{auc_violin}
\alias{auc_violin}
\title{Plot extracted AUCs by parameter settings}
\usage{
auc_violin(
  df.lis,
  xlab,
  ylab = "AUC",
  nrow = 3,
  title = NULL,
  key.title = NULL,
  lgd.key.size = 0.05
)
}
\arguments{
\item{df.lis}{The nested list of extracted aucs returned by \code{auc_stat}.}

\item{xlab, ylab}{The x and y axis labels in the violin plots.}

\item{nrow}{The numbers of rows of all the violin plots.}

\item{title}{The title of composite violin plots.}

\item{key.title}{The title of legend.}

\item{lgd.key.size}{The size of legend keys.}
}
\value{
An object of ggplot.
}
\description{
In coclustering optimization, visualize extracted AUCs by each parameter settings in violin plots.
}
\examples{
See function "coclus_opt" by calling "?coclus_opt".
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
Baptiste Auguie (2017). gridExtra: Miscellaneous Functions for "Grid" Graphics. R package version 2.3. https://CRAN.R-project.org/package=gridExtra
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
