% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclus_opt.R
\name{coclus_multi}
\alias{coclus_multi}
\title{Optimizing coclustering process with two levels of parallelizations available.}
\usage{
coclus_multi(
  wk.dir,
  sc.dim.min = 10,
  max.dim = 50,
  dimred = c("PCA", "UMAP"),
  graph.meth = c("knn", "snn"),
  sim = seq(0.2, 0.8, by = 0.1),
  sim.p = seq(0.2, 0.8, by = 0.1),
  dim = seq(5, 40, by = 1),
  df.match,
  sim.meth = "spearman",
  df.spd = NULL,
  batch.par = NULL,
  multi.core.par = MulticoreParam(workers = 1, RNGseed = NULL, stop.on.error = FALSE,
    log = TRUE, logdir = file.path(wk.dir, "multi_core_log")),
  verbose = TRUE
)
}
\arguments{
\item{wk.dir}{Working directory. Must be the same with that in \code{filter_iter}, since the filtered data will be used automatically.}

\item{dimred}{Dimensionality redeuction methods to optimize: \code{c('PCA', 'UMAP')}.}

\item{graph.meth}{Graph-building methods in coclustering: \code{c('knn', 'snn')} (see \code{\link[scran]{buildKNNGraph}}), \code{\link[scran]{buildSNNGraph}}) respectively). The clusters are detected by first creating a nearest neighbor graph using \code{snn} or \code{knn} then partitioning the graph.}

\item{sim, sim.p}{Used when refining cell clusters. Both are numeric scalars, ranging from 0 to 1. \code{sim} is a similarity (Spearman or Pearson correlation coefficient) cutoff between cells and \code{sim.p} is a proportion cutoff. In a certain cell cluster, cells having similarity >= \code{sim} with other cells in the same cluster at proportion >= \code{sim.p} would remain. Otherwise, they are discarded. The default of both is \code{seq(0.2, 0.8, by=0.1)} and can be customized.}

\item{dim}{Number of principle components (PCs, equivalent to genes) in combined bulk and single cell data. Used as the minimum number of PCs to retain in \code{\link[scran]{denoisePCA}} when coclustering bulk and single cells. The default is \code{seq(5, 40, by=1)}, and can be customized.}

\item{df.match}{The \code{data.frame} specifying matching between cells and true bulk.}

\item{sim.meth}{Method to calculate similarities between bulk and cells in each cocluster when assigning bulk to cells. \code{spearman} (default) or \code{pearson}.}

\item{df.spd}{A \code{data.frame} of \code{sim}, \code{sim.p}, \code{dim} sets, where one refers to a set of \code{sim}, \code{sim.p}, \code{dim} and the column names are \code{sim}, \code{sim.p}, \code{dim} respectively. For example, \code{data.frame(sim=0.2, sim.p=0.5, dim=10)}.}

\item{batch.par}{The parameters for first-level parallelization through a cluster scheduler such as SLURM. See \code{\link[BioParallel]{BatchtoolsParam}}. If \code{NULL} (default), the first-level parallelization is skipped.}

\item{multi.core.par}{The parameters for second-level parallelization. See \code{\link[BioParallel]{MulticoreParam}}.}

\item{verbose}{Logical. If \code{TRUE} (default), intermediate messages are printed.}
}
\value{
A list of normalized single cell and bulk data.
}
\description{
Optimizing coclustering process with two levels of parallelizations available.
}
\references{
Li, S., Yamada, M., Han, X., Ohler, U., and Benfey, P. N. (November, 2016) High-Resolution Expression Map of the Arabidopsis Root Reveals Alternative Splicing and lincRNA Regulation. Dev. Cell, 39(4), 508–522.
Shahan, R., Hsu, C.-W., Nolan, T. M., Cole, B. J., Taylor, I. W., Vlot, A. H. C., Benfey, P. N., and Ohler, U. (June, 2020) A single cell Arabidopsis root atlas reveals developmental trajectories in wild type and cell identity mutants.
Morgan M, Wang J, Obenchain V, Lang M, Thompson R, Turaga N (2021). BiocParallel: Bioconductor facilities for parallel evaluation. R package version 1.28.3, https://github.com/Bioconductor/BiocParallel.
Yoo, A. B., Jette, M. A., and Grondona, M. (2003) SLURM: Simple Linux Utility for Resource Management. In Job Scheduling Strategies for Parallel Processing Springer Berlin Heidelberg pp. 44–60.
McCarthy DJ, Campbell KR, Lun ATL, Willis QF (2017). “Scater: pre-processing, quality control, normalisation and visualisation of single-cell RNA-seq data in R.” Bioinformatics, 33, 1179-1186. doi: 10.1093/bioinformatics/btw777
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
