% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclus_roc.R
\name{coclus_roc}
\alias{coclus_roc}
\title{Co-cluster bulk and single cell data Calculate ROC/AUC for the combined bulk and single cell data}
\usage{
coclus_roc(
  bulk,
  cell.refined,
  df.match,
  min.dim = 10,
  max.dim = 50,
  graph.meth = "snn",
  dimred = "PCA",
  sim.meth = "spearman",
  seed = 1000
)
}
\arguments{
\item{bulk}{The normalized and filtered bulk data in form of \code{SingleCellExperiment}, \code{matrix} (log2-scale), or \code{data.frame} (log2-scale).}

\item{cell.refined}{The refined cell data in form of \code{SingleCellExperiment}, which is returned by \code{refine_cluster}.}

\item{df.match}{The \code{data.frame} specifying matching between cells and true bulk.}

\item{min.dim}{Integer scalars specifying the minimum (\code{min.dim}) and maximum (\code{max.dim}) number of (principle components) PCs to retain respectively in \code{\link[scran]{denoisePCA}}. The default is \code{min.dim=5}, \code{max.dim=50}.}

\item{max.dim}{Integer scalars specifying the minimum (\code{min.dim}) and maximum (\code{max.dim}) number of (principle components) PCs to retain respectively in \code{\link[scran]{denoisePCA}}. The default is \code{min.dim=5}, \code{max.dim=50}.}

\item{graph.meth}{Method to build a nearest-neighbor graph, \code{snn} (see \code{\link[scran]{buildSNNGraph}}) or \code{knn} (default, see \code{\link[scran]{buildKNNGraph}}). The clusters are detected by first creating a nearest neighbor graph using \code{snn} or \code{kn} then partitioning the graph.}

\item{dimred}{A string of \code{PCA} (default) or \code{UMAP} specifying which reduced dimensionality to use in creating a nearest neighbor graph. Internally, before building a nearest neighbor graph the data dimensionalities are reduced by PCA and UMAP respectively.}

\item{sim.meth}{Method to calculate similarities between bulk and cells in each cocluster when assigning bulk to cells. \code{spearman} (default) or \code{pearson}.}

\item{seed}{An integer for setting the seed of R's random number generator, which is useful to reproduce results.}
}
\value{
A list of \code{roc} object and the data frame to create the \code{roc}.
}
\description{
Co-cluster bulk and refined single cell data and assign bulk to single cells. Since the identities of bulk tissues and single cells are labeled, ROC/AUC are calculated to evaluate the co-clustering performance.
}
\examples{
See function "cocluster" by running "?cocluster".
}
\references{
Amezquita R, Lun A, Becht E, Carey V, Carpp L, Geistlinger L, Marini F, Rue-Albrecht K, Risso D, Soneson C, Waldron L, Pages H, Smith M, Huber W, Morgan M, Gottardo R, Hicks S (2020). “Orchestrating single-cell analysis with Bioconductor.” Nature Methods,    17, 137–145. https://www.nature.com/articles/s41592-019-0654-x.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
