% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_svg.R
\name{color_svg}
\alias{color_svg}
\title{Coloring spatial features in aSVG files.}
\usage{
color_svg(input, out.dir)
}
\arguments{
\item{input}{The output returned by \link{shm} or \link{covis}.}

\item{out.dir}{The directory path where the colored aSVG file will be saved.}
}
\value{
Nothing is returned.
}
\description{
This function copies the original aSVG file into a specified directory, and colors the spatial features in the new aSVG file with heat colors in spatial heatmaps.
}
\examples{

# Read the aSVG file.
svg.hum.pa <- system.file("extdata/shinyApp/data", 'homo_sapiens.brain.svg', 
package="spatialHeatmap")
svg.hum <- read_svg(svg.hum.pa)
# Attributes of spatial features.
feature.hum <- attribute(svg.hum)[[1]]
set.seed(20) # To obtain reproducible results, a fixed seed is set.
# Spatial features.
unique(feature.hum$feature)[1:10]
# Create a random numeric vector.
my_vec <- setNames(sample(1:100, 4), c('substantia.nigra', 'putamen', 
'prefrontal.cortex', 'notMapped'))
my_vec
# Plot spatial heatmaps with the imported aSVG and random vector.
dat.quick <- SPHM(svg=svg.hum, bulk=my_vec)
shm.res <- shm(data=dat.quick, ID='testing', ncol=1, sub.title.size=20, legend.nrow=3,
bar.width=0.1)
# Color the spatial features in the original aSVG file with heat colors in the spatial
# heatmap. The output aSVG file is saved in a temporary directory.
color_svg(input=shm.res, out.dir=tempdir(check = TRUE))
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
