% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_methods.R
\docType{methods}
\name{coordMethods}
\alias{coordMethods}
\alias{coordinate}
\alias{attribute}
\alias{dimension}
\alias{svg}
\alias{raster}
\alias{cmb}
\alias{coordinate,coord-method}
\alias{coordinate<-,coord-method}
\alias{attribute,coord-method}
\alias{attribute<-,coord-method}
\alias{dimension,coord-method}
\alias{dimension<-,coord-method}
\alias{raster,coord-method}
\alias{raster<-,coord-method}
\alias{svg,coord-method}
\alias{svg<-,coord-method}
\alias{[,coord,ANY,ANY,ANY-method}
\alias{[<-,coord,ANY,ANY,ANY-method}
\alias{length,coord-method}
\alias{names,coord-method}
\alias{names<-,coord-method}
\alias{cmb,coord,coord-method}
\title{Methods for S4 class \code{coord}}
\usage{
\S4method{coordinate}{coord}(x)

\S4method{coordinate}{coord}(x) <- value

\S4method{attribute}{coord}(x)

\S4method{attribute}{coord}(x) <- value

\S4method{dimension}{coord}(x)

\S4method{dimension}{coord}(x) <- value

\S4method{raster}{coord}(x)

\S4method{raster}{coord}(x) <- value

\S4method{svg}{coord}(x)

\S4method{svg}{coord}(x) <- value

\S4method{[}{coord,ANY,ANY,ANY}(x, i, j)

\S4method{[}{coord,ANY,ANY,ANY}(x, i) <- value

\S4method{length}{coord}(x)

\S4method{names}{coord}(x)

\S4method{names}{coord}(x) <- value

\S4method{cmb}{coord,coord}(x, y)
}
\description{
These are methods for subsetting, getting, setting, or combining \linkS4class{coord} objects.
}
\section{Main methods}{

In the following code snippets, \code{cordn} is a \linkS4class{coord} object.
\describe{
\item{\code{cordn[i]}, \code{cordn[i, ]}}{
  Subsetting the ith aSVG instance.
}
\item{\code{cordn[i] <- cordn.new}}{
  Replacing the ith aSVG instance in \code{cordn} with a new \code{cordn} object \code{cordn.new}.
}
\item{\code{cordn[, j]}}{
  Subsetting the jth slot of all aSVG instances.
}
\item{\code{cordn[, 'coordinate']}, \code{coordinate(cordn)}}{
  Subsetting the \code{coordinate} slot that contains coordinates of all aSVG instances.
}
\item{\code{length(cordn)}}{
  Number of all aSVG instances.
}
\item{\code{names(cordn)}, \code{names(cordn)[1] <- 'newName'}}{
  Names of all aSVG instances, rename the first aSVG instance.
}
\item{\code{cbm(cordn1, cordn2)}}{
  Combining two aSVG instances.
}
}
}

\examples{
# Create the first aSVG instance. 
svg.pa1 <- system.file('extdata/shinyApp/example/maize_leaf_shm1.svg',
package='spatialHeatmap')
svg1 <- parse_svg(svg.path=c(svg.pa1)); names(svg1); length(svg1); slotNames(svg1)
# Create the second aSVG instance. 
svg.pa2 <- system.file('extdata/shinyApp/example/maize_leaf_shm2.svg',
package='spatialHeatmap')
svg2 <- parse_svg(svg.path=c(svg.pa2)); names(svg2); length(svg2)
# Combine these two instances.
svg3 <- cmb(svg1, svg2); names(svg3); length(svg3)
# The first aSVG instance
svg3[1]
# Coordinates of the first aSVG instance 
svg3[, 'coordinate'][1]; coordinate(svg3)[1]
# Extract slots from "svg3" into a list and create a new "coord" object.
lis <- list(cordn=coordinate(svg3), attrb=attribute(svg3), svg=svg(svg3))
new.svgs <- coord(coordinate=lis$cordn, attribute=lis$attrb, svg=lis$svg)
# Change aSVG instance names.
names(new.svgs) <- c('aSVG1', 'aSVG2'); names(new.svgs)
# Replace an instance.
svg3[2] <- new.svgs[2]
# Replace a slot content.
coordinate(svg3)[[1]] <- coordinate(new.svgs)[[1]]
}
\seealso{
\code{\link{coord}}: creating \code{coord} objects.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
