% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_table_data.R
\docType{data}
\name{deg.table}
\alias{deg.table}
\title{A table of differentialy-expressed genes (DEGs) detected by different methods}
\format{
A table.
}
\source{
\href{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6769/}{ExpressionAtlas E-MTAB-6769}
}
\usage{
data(deg.table)
}
\description{
A table of up- and down-DEGs detected by different methods such as edgeR, limma, DEseq2.
}
\examples{
data(deg.table)
deg.table[1:2, ]
}
\references{
Cardoso-Moreira, Margarida, Jean Halbert, Delphine Valloton, Britta Velten, Chunyan Chen, Yi Shao, Angélica Liechti, et al. 2019. “Gene Expression Across Mammalian Organ Development.” Nature 571 (7766): 505–9
}
\keyword{datasets}
