% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter.data}
\alias{filter.data}
\title{Filter the Data Matrix}
\usage{
filter.data(data, sep, isRowGene, pOA = c(0, 0), CV = c(0, 10000),
  dir = "./")
}
\arguments{
\item{data}{The path of the data matrix. In the gene expression matrix, the row and column names are usually gene IDs and sample/conditions, respectively. The sample/condition names MUST be fomatted this way: a sample name is followed by double underscore then the condition, such as “stele__140mM_48h" (Mustroph et al. 2009), where stele is the sample and 140mM_48h is the condition. One column of meta data (e.g. gene annotation) can also be included in parallel with sample/condition at the end. In the column names of sample/condition and meta data, only letters, digits, single underscore, dots are allowed. Not all samples in the matrix necessarily need to be present in the svg image, and vice versa. Only samples present in the svg image are recognised and coloured. The rows and columns can be swaped and the function identifies gene IDs by "isRowGen". \cr E.g.: system.file("extdata/example", "root_expr_ann_row_gen.txt", package = "spatialHeatmap").}

\item{sep}{The seprator of the data matrix, e.g. ",", "\\t", ";".}

\item{isRowGene}{If row names are gene IDs it is "TRUE", otherwise it is "FALSE". If the data matrix is not gene expression, this argument is set according to if the row is equivalent to gene ID.}

\item{pOA}{It specifies parameters of the filter function "pOverA" from the package "genefilter" (Gentleman et al. 2018). It filters genes according to the proportion "p" of samples where the expression values exceeding a threshold "A". The input is a two-component vector, where the first one is the "p" and the second one is "A". The default is c(0, 0), which means no filter is applied. \cr E.g. c(0.1, 2) means genes whose expression values over 2 in at least 10\% of all samples are kept.}

\item{CV}{It specifies parameters of the filter function "cv" from the package "genefilter" (Gentleman et al. 2018), which filters genes according to the coefficient of variation (CV). The input is a two-component vector, specifying the CV range. The default is c(1, 10000), where the range is set from very samll (0) to very large (10000) so as to not apply filtering. \cr E.g. c(0.1, 5) means genes with CV between 0.1 and 5 are kept.}

\item{dir}{The location where the directory "local_mode_result/" is created automatically to save the processed data matrix. This argument should be the same with that from the function "adj.mod" so that the files "adj.txt" and "mod.txt" generated by "adj.mod" are saved in the same place.}
}
\value{
A data matrix filtered by "pOverA" and "cv" from the R package "genefilter", which is saved in the directory "local_mode_result". It is a "\\t" separated "txt" file and row names (gene IDs) will be used to display the spatial heatmaps.
}
\description{
It is designed to filter out the noise included in the gene expression matrix. There are two internal filter functions "pOverA" and "cv", which are from the package "genefilter". The processed matrix (processed_data.txt) is saved in the directory "local_mode_result", which is automatically created in the location specified by "dir". \cr When using the web-browser based Shiny app, the expression matrix can also be processed with this function and then uploaded to "Step 2: upload a gene expression file".
}
\examples{
data.path <- system.file("extdata/example", "root_expr_ann_row_gen.txt", package = "spatialHeatmap")
exp <- filter.data(data=data.path, sep="\\t", isRowGene=TRUE, pOA=c(0, 0), CV=c(0.1, 10000), dir="./")
}
\references{
R. Gentleman, V. Carey, W. Huber and F. Hahne (2017). genefilter: genefilter: methods for filtering genes from high-throughput experiments. R package version 1.58.1 \cr Matt Dowle and Arun Srinivasan (2017). data.table: Extension of `data.frame`. R package version 1.10.4. https://CRAN.R-project.org/package=data.table 
Mustroph, Angelika, M Eugenia Zanetti, Charles J H Jang, Hans E Holtan, Peter P Repetti, David W Galbraith, Thomas Girke, and Julia Bailey-Serres. 2009. “Profiling Translatomes of Discrete Cell Populations Resolves Altered Cellular Priorities During Hypoxia in Arabidopsis.” Proc Natl Acad Sci U S A 106 (44): 18843–8
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
