% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cell.R
\name{filter_cell}
\alias{filter_cell}
\title{Filter single cell data separately in a list}
\usage{
filter_cell(
  lis,
  bulk = NULL,
  gen.rm = NULL,
  min.cnt = 1,
  p.in.cell = 0.4,
  p.in.gen = 0.2,
  verbose = TRUE
)
}
\arguments{
\item{lis}{A named list of single cell data in form of \code{data.frame}, \code{SingleCellExperiment}, or \code{SummarizedExperiment}.}

\item{bulk}{The bulk data in form of \code{data.frame} or \code{SummarizedExperiment}. They are only used to obtain common genes with all single cell data and not filtered. The default is \code{NULL}.}

\item{gen.rm}{A regular expression of gene identifiers in single cell data to remove before filtering. E.g. mitochondrial, chloroplast and protoplasting-induced genes (\code{^ATCG|^ATCG}). The default is \code{NULL}.}

\item{min.cnt}{The min count of gene expression. The default is \code{1}.}

\item{p.in.cell}{The min proportion of counts above \code{min.cnt} in a cell. The default is \code{0.4}.}

\item{verbose}{Logical. If \code{TRUE} (default), intermediate messages are printed.}

\item{p.in.gene}{The min proportion of counts above \code{min.cnt} in a gene. The default is \code{0.2}.}
}
\value{
A list of filtered single cell data and bulk data, which have common genes.
}
\description{
Filter single cell data separately in a list and take overlap genes between all single cell and bulk data. The bulk data are not filtered and are only used to obtain overlap genes.
}
\examples{
# Example bulk data of mouse brain for coclustering (Vacher et al 2021). 
blk.mus.pa <- system.file("extdata/shinyApp/example", "bulk_mouse_cocluster.txt", package="spatialHeatmap")
blk.mus <- as.matrix(read.table(blk.mus.pa, header=TRUE, row.names=1, sep='\t', check.names=FALSE))  
blk.mus[1:3, 1:5] 
# Example single cell data for coclustering (Ortiz et al 2020).
sc.mus.pa <- system.file("extdata/shinyApp/example", "cell_mouse_cocluster.txt", package="spatialHeatmap")  
sc.mus <- as.matrix(read.table(sc.mus.pa, header=TRUE, row.names=1, sep='\t', check.names=FALSE)) 
sc.mus[1:3, 1:5]  
# Initial filtering. 
blk.mus <- filter_data(data=blk.mus, sam.factor=NULL, con.factor=NULL, pOA=c(0.1, 5), CV=c(0.2, 100), dir=NULL)
dim(blk.mus) 
mus.lis <- filter_cell(lis=list(sc.mus=sc.mus), bulk=blk.mus, gen.rm=NULL, min.cnt=1, p.in.cell=0.5, p.in.gen=0.1)
}
\references{
Martin Morgan, Valerie Obenchain, Jim Hester and Hervé Pagès (2021). SummarizedExperiment: SummarizedExperiment container. R package version 1.24.0. https://bioconductor.org/packages/SummarizedExperiment
Amezquita R, Lun A, Becht E, Carey V, Carpp L, Geistlinger L, Marini F, Rue-Albrecht K, Risso D, Soneson C, Waldron L, Pages H, Smith M, Huber W, Morgan M, Gottardo R, Hicks S (2020). "Orchestrating single-cell analysis with Bioconductor." _Nature Methods_, *17*, 137-145. <URL: https://www.nature.com/articles/s41592-019-0654-x>
Douglas Bates and Martin Maechler (2021). Matrix: Sparse and Dense Matrix Classes and Methods. R package version 1.4-0. https://CRAN.R-project.org/package=Matrix
Vacher CM, Lacaille H, O'Reilly JJ, Salzbank J et al. Placental endocrine function shapes cerebellar development and social     behavior. Nat Neurosci 2021 Oct;24(10):1392-1401. PMID: 34400844.                                                                  
Ortiz C, Navarro JF, Jurek A, Märtin A et al. Molecular atlas of the adult mouse brain. Sci Adv 2020 Jun;6(26):eabb3446. PMID:  32637622
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
