% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_iter.R
\name{filter_iter}
\alias{filter_iter}
\title{Iteratively filter bulk and single data according to parameters in a data frame.}
\usage{
filter_iter(bulk, cell.lis, df.par.fil, gen.rm = NULL, norm.meth, wk.dir)
}
\arguments{
\item{bulk}{Normalized bulk data at log2-scale returned by \code{norm_multi}.}

\item{cell.lis}{Normalized single cell data at log2-scale in a named list returned by \code{norm_multi}.}

\item{df.par.fil}{A \code{data.frame} of filtering parameter settings that are passed to \code{filter_data} and \code{filter_cell} respectively. E.g. \code{df.par.fil <- data.frame(p=c(0.2, 0.3), A=rep(1, 4), cv1=c(0.2, 0.3), cv2=rep(100, 4), min.cnt=rep(1, 4), p.in.cell=c(0.25, 0.3), p.in.gen=c(0.05, 0.1))}.}

\item{gen.rm}{A regular expression of gene identifiers in single cell data to remove before filtering. E.g. mitochondrial, chloroplast and protoplasting-induced genes (\code{^ATCG|^ATCG}). The default is \code{NULL}.}

\item{norm.meth}{Methods used to normalize bulk and single cell data. One of \code{fct} and \code{cpm}, standing for \code{\link[scran]{computeSumFactors}} only and further normalized by counts per million (cpm) respectively. No actual normalization is performed, only used in file names when saving filtered results.}

\item{wk.dir}{The work directory where filtered data are saved in ".rds" files by \code{saveRDS}.}
}
\value{
Filtered data are save in \code{wk.dir}.
}
\description{
In secondary filtering of coclustering optimization, iteratively filter normalized bulk and single data (in a list) according to parameter combinations in a data frame.
}
\examples{
See function "coclus_opt" by calling "?coclus_opt".   
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
