% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_setting.R
\name{final_setting}
\alias{final_setting}
\title{Extracting combinations of optimal settings across data sets}
\usage{
final_setting(auc.lis, auc.min = 0)
}
\arguments{
\item{auc.lis}{A list of AUCs from each single cell dataset that is returned by \code{auc_by_dataset}.}

\item{auc.min}{An AUC value cutoff. Only AUCs above this cutoff are considered.}
}
\value{
An object of ggplot.
}
\description{
After coclustering optimization, visualize AUCs generated by optimal parameter settings on validating data sets.
}
\examples{
# Create parameter settings and corresponding aucs.
df.auc <- data.frame(auc=runif(n=5, min=1e-12, max=.99))
ran <- seq(0.2, 0.8, 0.1)
df.spd <- data.frame(sim=sample(ran, 5, replace=FALSE), sim.p=sample(ran, 5, replace=FALSE), dim=sample(seq(5, 40, 1), 5, replace=FALSE))
df.spd$spd.set <- paste0('s', df.spd$sim, 'p', df.spd$sim.p, 'd', df.spd$dim)

df.auc <- cbind(df.spd, df.auc)
# Plot the aucs.
auc_bar(df.auc, auc='auc', thr=0.5) 
# See package vignette by calling "browseVignettes('spatialHeatmap')".
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
