% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_2lgd.R
\name{gg_2lgd}
\alias{gg_2lgd}
\title{Add Secondary Legend}
\usage{
gg_2lgd(
  gg.all,
  sam.dat,
  tis.trans,
  position.2nd = "bottom",
  legend.nrow.2nd = NULL,
  legend.ncol.2nd = NULL,
  legend.key.size.2nd = 0.03,
  add.feature.2nd = FALSE,
  legend.text.size.2nd = 10,
  angle.text.key.2nd = 0,
  position.text.key.2nd = "right"
)
}
\arguments{
\item{gg.all}{A list of spatial heatmaps of ggplot.}

\item{sam.dat}{A vector of samples in the data matrix.}

\item{tis.trans}{A vector of tissues to be transparent.}

\item{position.2nd}{The position of the secondary legend. One of "top", "right", "bottom", "left", or a two-component numeric vector. The default is "bottom". Applies to the static image and video.}

\item{legend.nrow.2nd}{An integer of rows of the secondary legend keys. Applies to the static image and video.}

\item{legend.ncol.2nd}{An integer of columns of the secondary legend keys. Applies to the static image and video.}

\item{legend.key.size.2nd}{A numeric of legend key size. The default is 0.03. Applies to the static image and video.}

\item{add.feature.2nd}{Logical TRUE or FALSE. Add feature identifiers to the secondary legend or not. The default is FALSE. Applies to the static image.}

\item{legend.text.size.2nd}{A numeric of the secondary legend text size. The default is 10. Applies to the static image and video.}

\item{angle.text.key.2nd}{A value of angle of key text in the secondary legend. Default is 0. Applies to the static image and video.}

\item{position.text.key.2nd}{The position of key text in the secondary legend, one of "top", "right", "bottom", "left". Default is "right". Applies to the static image and video.}
}
\value{
A list of ggplots.
}
\description{
Add Secondary Legend
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
