% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_lgd.R
\name{gg_lgd}
\alias{gg_lgd}
\title{Adjust Legend Key Size and Rows in Ggplot.}
\usage{
gg_lgd(
  gg.all,
  size.key = 0.02,
  size.text = 8,
  row = 2,
  sam.dat,
  tis.trans = NULL
)
}
\arguments{
\item{gg.all}{A list of spatial heatmaps of ggplot.}

\item{size.key}{A numeric of legend key size. If \code{size.text} is NULL, it also applies to legend text size.}

\item{size.text}{A numeric of legend text size.}

\item{row}{An integer of rows in legend key.}

\item{sam.dat}{A vector of samples in the data matrix.}

\item{tis.trans}{A vector of tissues to be transparent.}
}
\value{
A list of ggplots.
}
\description{
Adjust Legend Key Size and Rows in Ggplot.
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
