% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_gg.R
\name{grob_gg}
\alias{grob_gg}
\title{Separate SHMs of grobs from ggplot}
\usage{
grob_gg(gs)
}
\arguments{
\item{gs}{A list of SHMs of grobs, legend plot, and SHMs of ggplot, returned by \code{\link{grob_list}}.}
}
\value{
A list of grob SHMs and ggplot SHMs.
}
\description{
Separate SHMs of grobs from ggplot. Different SHMs of same 'gene_condition' are indexed with suffixed of '_1', '_2', ...
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
