% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_list.R
\name{grob_list}
\alias{grob_list}
\title{Draw Each Spatial Heatmap and Convert Them to Ggplot2 Plot Grob}
\usage{
grob_list(
  gene,
  con.na = TRUE,
  geneV,
  coord,
  ID,
  cols,
  tis.path,
  tis.trans = NULL,
  sub.title.size,
  sam.legend = "identical",
  legend.col,
  legend.ncol = NULL,
  legend.nrow = NULL,
  legend.position = "bottom",
  legend.direction = NULL,
  legend.key.size = 0.02,
  legend.text.size = 12,
  legend.plot.title = NULL,
  legend.plot.title.size = 11,
  line.size = 0.2,
  line.color = "grey70",
  mar.lb = NULL,
  ...
)
}
\arguments{
\item{gene}{The gene expession matrix, where rows are genes and columns are tissue/conditions.}

\item{con.na}{Logical, TRUE or FALSE. Default is TRUE, meaning conditions are available.}

\item{geneV}{The gene expression values used to construct the colour bar.}

\item{coord}{The coordidates extracted from the SVG file.}

\item{ID}{All gene ids selected after the App is launched.}

\item{cols}{All the color codes used to construct the color bar.}

\item{tis.path}{All the tissues/paths extracted from the SVG.}

\item{tis.trans}{A character vector of tissue/spatial feature identifiers that will be set transparent. \emph{E.g} c("brain", "heart"). This argument is used when target features are covered by  overlapping features and the latter should be transparent.}

\item{sub.title.size}{A numeric of the subtitle font size of each individual spatial heatmap. The default is 11.}

\item{sam.legend}{One of "identical", "all", or a character vector of tissue/spatial feature identifiers from the aSVG file. The default is "identical" and all the identical/matching tissues/spatial features between the data and aSVG file are indicated in the legend plot. If "all", all tissues/spatial features in the aSVG are shown. If a vector, only the tissues/spatial features in the vector are shown.}

\item{legend.col}{A character vector of colors for the keys in the legend plot. The lenght must be equal to the number of target samples shown in the legend.}

\item{legend.ncol}{An integer of the total columns of keys in the legend plot. The default is NULL. If both \code{legend.ncol} and \code{legend.nrow} are used, the product of the two arguments should be equal or larger than the total number of shown spatial features.}

\item{legend.nrow}{An integer of the total rows of keys in the legend plot. The default is NULL. It is only applicable to the legend plot. If both \code{legend.ncol} and \code{legend.nrow} are used, the product of the two arguments should be equal or larger than the total number of matching spatial features.}

\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}

\item{legend.direction}{layout of items in legends ("horizontal" or
"vertical")}

\item{legend.key.size}{A numeric of the legend key size ("npc"), applicable to the legend plot. The default is 0.02.}

\item{legend.text.size}{A numeric of the legend label size, applicable to the legend plot. The default is 12.}

\item{legend.plot.title}{The title of the legend plot. The default is NULL.}

\item{legend.plot.title.size}{The title size of the legend plot. The default is 11.}

\item{line.size}{A numeric of the shape outline size. Default is 0.2.}

\item{line.color}{A character of the shape outline color. Default is "grey70".}

\item{mar.lb}{A two-component numeric vector. The first and second numeric is left/right and bottom/top margin (npc) respectively.}

\item{...}{Other arguments passed to \code{\link[ggplot2]{ggplot}}.}
}
\value{
A nested list of spatial heatmaps of ggplot2 plot grob, spatial heatmaps of ggplot, and legend plot of ggplot.
}
\description{
Draw Each Spatial Heatmap and Convert Them to Ggplot2 Plot Grob
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016. \cr R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. RL https://www.R-project.org/ \cr Mustroph, Angelika, M Eugenia Zanetti, Charles J H Jang, Hans E Holtan, Peter P Repetti, David W Galbraith, Thomas Girke, and Julia Bailey-Serres. 2009. “Profiling Translatomes of Discrete Cell Populations Resolves Altered Cellular Priorities During Hypoxia in Arabidopsis.” Proc Natl Acad Sci U S A 106 (44): 18843–8
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
