% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lis_deg_up_down_data.R
\docType{data}
\name{lis.deg.up.down}
\alias{lis.deg.up.down}
\title{A nested list of differentialy-expressed genes (DEGs) detected by different methods}
\format{
A nested list.
}
\source{
\href{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6769/}{ExpressionAtlas E-MTAB-6769}
}
\usage{
data(lis.deg.up.down)
}
\description{
A nested list of up- and down-DEGs detected by different methods such as edgeR, limma, DEseq2.
}
\examples{
data(lis.deg.up.down)
lis.deg.up.down$up.lis$edgeR.up[1:5]
}
\references{
Cardoso-Moreira, Margarida, Jean Halbert, Delphine Valloton, Britta Velten, Chunyan Chen, Yi Shao, Angélica Liechti, et al. 2019. “Gene Expression Across Mammalian Organ Development.” Nature 571 (7766): 505–9
}
\keyword{datasets}
