% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_cluster.R
\name{manual_cluster}
\alias{manual_cluster}
\title{Add manual cell cluster labels to SingleCellExperiment}
\usage{
manual_cluster(sce, df.clus)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment}.}

\item{df.clus}{A \code{data.frame} of manually created single cell clusters. The \code{cell} column contains cell identifiers that are present in \code{rownames(colData(sce))} and the \code{cluster} column contains the manually created cluster labels.}
}
\value{
An object of \code{SingleCellExperiment}.
}
\description{
Add manually created cell cluster labels in a \code{data.frame} to \code{SingleCellExperiment}.
}
\examples{
set.seed(10)
# Read single cell data.
sce.pa <- system.file("extdata/shinyApp/example", "sce_manual_mouse.rds", package="spatialHeatmap")
sce <- readRDS(sce.pa)
# Quality control, normalization, dimensionality reduction on the single cell data.
sce.dimred <- process_cell_meta(sce.manual, qc.metric=list(subsets=list(Mt=rowData(sce.manual)$featureType=='mito'), threshold=1))
# Read manual cell cluster labels.
manual.clus.mus.sc.pa <- system.file("extdata/shinyApp/example", "manual_cluster_mouse_brain.txt", package="spatialHeatmap") 
manual.clus.mus.sc <- read.table(manual.clus.mus.sc.pa, header=TRUE, sep='\t')
# Include manual cell cluster labels in "SingleCellExperiment".
sce.clus <- manual_cluster(sce=sce.dimred, df.clus=df.clus)
}
\references{
Morgan M, Obenchain V, Hester J, Pagès H (2022). SummarizedExperiment: SummarizedExperiment container. R package version 1.26.1, https://bioconductor.org/packages/SummarizedExperiment
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
