% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_heatmap.R
\name{matrix.heatmap}
\alias{matrix.heatmap}
\title{Matrix Heatmap}
\usage{
matrix.heatmap(geneID, expr.table, adj.mod, ds, scale)
}
\arguments{
\item{geneID}{A gene ID from the expression matrix.}

\item{expr.table}{The gene expression table, where rows are gene IDs and columns are samples. If annotation is included, it must be in the last column.}

\item{adj.mod}{The "adjacency matrix" and "modules" definitions resulting from the function "adj.mod".}

\item{ds}{The module identification sensitivity. See function "adj.mod" for details.}

\item{scale}{It specifies whether to scale the heatmap. There are three options: "row" means scale by row, "column" means scale by column, and "no" means no scale.}
}
\value{
An interactive matrix lauched on the web browser.
}
\description{
This function displays the gene network module containing the input gene as an interactive matrix heatmap, where the rows and columns are sorted by hierarchical clustering dendrograms and the input gene is tagged by a red rectangle. To explore the results, the matrix heatmap has several interactive features. For instance, users can zoom in and out by drawing a rectangle and by double clicking the image, respectively. Users can scale the expression values by gene or sample. \cr The network modules are identified at two alternative sensitivities levels (3, 2). From 3 to 2, the sensitivity decreases and results in less modules with larger sizes. The same module can also be displayed as an interactive network by the "network" function.
}
\examples{
data.path <- system.file("extdata/example", "root_expr_ann_row_gen.txt", package = "spatialHeatmap")
exp <- filter.data(data=data.path, sep="\\t", isRowGen=TRUE, c(0, 0), c(0.1, 10000), "processed_data")
adj_mod <- adj.mod(data=exp, type="signed", minSize=15)
\donttest{ matrix.heatmap("PSAC", exp, adj_mod, "2", "row") }
}
\references{
Andrie de Vries and Brian D. Ripley (2016). ggdendro: Create Dendrograms and Tree Diagrams Using 'ggplot2'. R package version 0.1-20. https://CRAN.R-project.org/package=ggdendro \cr H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016. \cr Carson Sievert (2018) plotly for R. https://plotly-book.cpsievert.me \cr Langfelder P and Horvath S, WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics 2008, 9:559 doi:10.1186/1471-2105-9-559 \cr Peter Langfelder, Steve Horvath (2012). Fast R Functions for Robust Correlations and Hierarchical Clustering. Journal of Statistical Software, 46(11), 1-17. URL http://www.jstatsoft.org/v46/i11/ \cr R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/ \cr Carson Sievert (2018) plotly for R. https://plotly-book.cpsievert.me
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
