% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_auc_bar.R
\name{mean_auc_bar}
\alias{mean_auc_bar}
\title{Plot mean of extracted AUCs by parameter settings}
\usage{
mean_auc_bar(
  df.auc,
  parameter = "parameter",
  auc.thr = "auc.thr",
  mean = "mean",
  bar.width = 0.8,
  title = NULL,
  key.title = NULL,
  lgd.key.size = 0.05
)
}
\arguments{
\item{df.auc}{The \code{data.frame} of mean AUCs in the nested list of extracted aucs returned by \code{auc_stat}.}

\item{parameter}{The coloumn name of parameters in \code{df.auc}.}

\item{auc.thr}{The coloumn name of AUC cutoffs in \code{df.auc}.}

\item{mean}{The coloumn name of mean AUCs in \code{df.auc}.}

\item{bar.width}{Width of a single bar.}

\item{title}{The title of composite violin plots.}

\item{key.title}{The title of legend.}

\item{lgd.key.size}{The size of legend keys.}
}
\value{
An object of ggplot.
}
\description{
In coclustering optimization, visualize means of extracted AUCs of each parameter settings by each AUC cutoff in bar plots.
}
\examples{
See function "coclus_opt" by calling "?coclus_opt".
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
