% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nod_lin.R
\name{nod_lin}
\alias{nod_lin}
\title{Calculate Gene Connectivity and Edge Adjacency}
\usage{
nod_lin(ds, lab, mods, adj, geneID, adj.min)
}
\arguments{
\item{ds}{Module detecting sensativity, "2" or "3".}

\item{lab}{An interger, which is the module the target gene belongs to.}

\item{mods}{A 2-column data frame, where each coloumn is a module assignmet.}

\item{adj}{The complete adjacency matrix.}

\item{geneID}{The target gene.}

\item{adj.min}{The minimum adjacency to keep for the edges.}
}
\value{
A 2-length list, containing the gene connectiviy and edge adjaceny respectively.
}
\description{
Calculate Gene Connectivity and Edge Adjacency
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
