% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_br.R
\name{path_br}
\alias{path_br}
\title{Break a Combined Path into a Group or Siblings}
\usage{
path_br(node, g = TRUE)
}
\arguments{
\item{node}{An object of class "xml_node" without children nodes.}

\item{g}{Logical, TRUE or FALSE. Default is TRUE. If TRUE the combined path is broken into a group. Otherwise, as siblings.}
}
\value{
Nothing is returned. The broken paths are updated in the root.
}
\description{
The combined paths introduces connecting lines between the last point of one polygon and the first point of next polygon in the spatial heatmaps. Therefore, they should be broken apart into a group or siblings. This function checks if the input node is a combined path internally.
}
\references{
Hadley Wickham, Jim Hester and Jeroen Ooms (2019). xml2: Parse XML. R package version 1.2.2. https://CRAN.R-project.org/package=xml2
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
