% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_br_all.R
\name{path_br_all}
\alias{path_br_all}
\title{Break Children in a Node into a Group or Siblings}
\usage{
path_br_all(node.parent)
}
\arguments{
\item{node.parent}{An object of class "xml_node" with children, i.e. the last two nodes in SVG file, outline and tissue layer respectively.}
}
\value{
Nothing is returned. The broken paths are updated in the root.
}
\description{
This function applies to the last two parent nodes in SVG file. Each child node in the input parent node is checked for combined path. The function \code{\link{path_br}} is used to break the combined paths.
}
\references{
Hadley Wickham, Jim Hester and Jeroen Ooms (2019). xml2: Parse XML. R package version 1.2.2. https://CRAN.R-project.org/package=xml2
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
