% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_para.R
\name{random_para}
\alias{random_para}
\title{Generate random combinations of parameter settings}
\usage{
random_para(
  fil.set,
  norm,
  dimred,
  graph.meth,
  sim = round(seq(0.2, 0.8, by = 0.1), 1),
  sim.p = round(seq(0.2, 0.8, by = 0.1), 1),
  dim = seq(5, 40, by = 1),
  df.spd.opt
)
}
\arguments{
\item{fil.set}{A character vector of filtering parameter set. E.g. \code{c('fil3', 'fil4')}.}

\item{norm}{A character vector of normalization methods. E.g. \code{c('cpm')}.}

\item{dimred}{A character vector of dimesionality reduction methods. E.g. \code{c('umap')}.}

\item{graph.meth}{A character vector of graph-building methods. E.g. \code{c('knn', 'snn')}.}

\item{df.spd.opt}{A \code{data.frame} of optimized \code{spd.set} settings. These settings are avoided in the output random settings.}
}
\value{
A \code{data.frame}.
}
\description{
In coclustering, generate random combinations of parameter settings for validating optimal settings.
}
\examples{

df.spd.opt <- data.frame(sim=c(0.2, 0.4, 0.3), sim.p=c(0.8, 0.6, 0.7), dim=c(12, 14, 13))
df.para.rdn <- random_para(fil.set=c('fil3', 'fil4'), norm='cpm', dimred='UMAP', graph.meth=c('knn', 'snn'), sim=round(seq(0.2, 0.8, by=0.1), 1), sim.p=round(seq(0.2, 0.8,by=0.1), 1), dim=seq(5, 40, by=1), df.spd.opt=df.spd.opt)
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
