% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdf5.R
\name{read_hdf5}
\alias{read_hdf5}
\title{Read Data from the Shiny App Database}
\usage{
read_hdf5(file, prefix)
}
\arguments{
\item{file}{The path of "data_shm.tar" generated by \code{write_hdf5}.}

\item{prefix}{A vector of data set identifiers such as \code{c('expr_arab', 'expr_chicken', 'df_pair')}. The vector elements must come from the "data" column in the pairing table that is made when calling \code{write_hdf5}.}
}
\value{
A list of data set and/or the pairing table.
}
\description{
This function is used to extract data from the Shiny App Database "data_shm.tar".
}
\examples{
# See examples in the function write_hdf5.
}
\references{
SummarizedExperiment: SummarizedExperiment container. R package version 1.10.1 \cr R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/
Hervé Pagès (2020). HDF5Array: HDF5 backend for DelayedArray objects. R package version 1.16.1.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
