% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dim.R
\name{reduce_dim}
\alias{reduce_dim}
\title{Reducing dimensionality in count data}
\usage{
reduce_dim(
  sce,
  choose.hvg = list(assay.type = "logcounts", block = NULL, more.var.args =
    list(transform = TRUE), top = 4000, more.choose.args = list(keep.ties = TRUE)),
  min.dim = 11,
  max.dim = 50,
  pca.arg = list(assay.type = "logcounts", block = NULL),
  umap.arg = list(num.dim = 2, reddim.type = "PCA"),
  tsne.arg = list(perplexity = 30, reddim.type = "PCA", output.name = "TSNE"),
  pca = FALSE
)
}
\arguments{
\item{sce}{Normalized single cell data in \code{SingleCellExperiment} returned by \code{norm_cell}. Alternative forms include \code{dgCMatrix}, \code{matrix}, \code{data.frame}.}

\item{choose.hvg}{Arguments in a named list passed to \code{\link[scrapper]{chooseRnaHvgs.se}}.}

\item{min.dim, max.dim}{Integer scalars specifying the minimum (\code{min.dim}) and maximum (\code{max.dim}) number of (principle components) PCs to retain respectively in \code{\link[scrapper]{runPca.se}}.}

\item{pca.arg}{Arguments in a named list passed to \code{\link[scrapper]{runPca.se}}.}

\item{umap.arg}{Arguments in a named list passed to \code{\link[scrapper]{runUmap.se}}.}

\item{tsne.arg}{Arguments in a named list passed to \code{\link[scrapper]{runTsne.se}}.}

\item{pca}{Logical, if \code{TRUE} only the data with reduced dimentionality of PCA is returned. The default is \code{FALSE}, and UMAP and TSNE are also returned.}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
A meta function for reducing dimensionality in count data.
}
\examples{
library(SingleCellExperiment)
set.seed(123)  # for reproducibility
# random counts: 1000 genes × 100 cells
counts_mat <- matrix(
  rpois(1000 * 100, lambda = 5), nrow = 1000, ncol = 100,
  dimnames = list(
    paste0("Gene", 1:1000),
    paste0("Cell", 1:100)
  )
)
sce <- SingleCellExperiment(
  assays = list(counts = counts_mat)
)
sce.qc = qc_cell(sce)
sce.norm <- norm_cell(sce.qc)
sce.dimred <- reduce_dim(sce.norm)
}
\references{
Amezquita R, Lun A, Becht E, Carey V, Carpp L, Geistlinger L, Marini F, Rue-Albrecht K, Risso D, Soneson C, Waldron L, Pages H, Smith M, Huber W, Morgan M, Gottardo R, Hicks S (2020). “Orchestrating single-cell analysis with Bioconductor.” Nature Methods,    17, 137–145. https://www.nature.com/articles/s41592-019-0654-x.
Lun ATL, Kancherla J (2023). "Powering single-cell analyses in the browser with WebAssembly." _Journal of Open Source Software_, *8*(89), 5603. doi:10.21105/joss.05603 <https://doi.org/10.21105/joss.05603>.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
