% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_gen_con.R
\name{sort_gen_con}
\alias{sort_gen_con}
\title{Sort "gene_condition" Names}
\usage{
sort_gen_con(ID.sel, na.all, con.all, by = "gene")
}
\arguments{
\item{ID.sel}{The vector of target genes.}

\item{na.all}{All "gene_condition" names.}

\item{con.all}{All condition names with suffixes of "_1", "_2", etc.}

\item{by}{"gene", "con", or "none". Sort conditions for each gene, vice versa, or no sortation.}
}
\value{
A vector of sorted "gene_condition" names.
}
\description{
Sort "gene_condition" Names
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
