% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_asg.R
\name{sub_asg}
\alias{sub_asg}
\title{Subset the bulk-cell assignments}
\usage{
sub_asg(res.lis, thr, true.only = TRUE)
}
\arguments{
\item{res.lis}{The result list of coclustering, which is the output of \code{tests} and comprises three slots \code{sce}, \code{roc.obj}, \code{df.roc}.}

\item{thr}{The threshold for subsetting bulk-cell assignments, which is a similarity value (Pearson's or Spearman's correlation coefficient) between bulk and cells. Only bulk-cell assignments with similarity values above the thresold would remain.}

\item{true.only}{Logical. If \code{TRUE}, only the true assignments are returned in the subset \code{SingleCellExperiment}. This argument affects the values for plottting SHMs.}
}
\value{
A \code{SingleCellExperiment} of remaining bulk-cell assignments.
}
\description{
Subset the bulk-cell assignments according to a threshold, which is a similarity value between bulk and cells.
}
