% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_na.R
\name{sub_na}
\alias{sub_na}
\title{Function to extract nearest genes.}
\usage{
sub_na(mat, ID, p = 0.3, n = NULL, v = NULL)
}
\arguments{
\item{mat}{The correlation or distance matrix.}

\item{ID}{A vector of target gene identifiers.}

\item{p}{The proportion of top assayed items with most similar expression profiles with the target items. Only items within this proportion are returned. Default is 0.3. It applies to each target item independently, and selected items of each target are returned together.}

\item{n}{An integer of top items with most similar expression profiles with the target items. Only items within this number are returned. Default is NULL. It applies to each target independently, and selected items of each target are returned together.}

\item{v}{A numeric of correlation (-1 to 1) or distance (>=0) threshold to select items with the most similar expression profiles with the target items. If \code{fun='cor'}, only items with correlation coefficient larger than \code{v} are returned. If \code{fun='dist'}, only items with distance less than \code{v} are returned. Default is NULL. It applies to each target independently, and selected items of each target are returned together.}
}
\value{
A vector of target genes and their nearest neighbors.
}
\description{
Given a vector of target gene identifiers, this function selects genes with similar expression profiles for every target gene. The selection is based on correlation matrix or distance matrix generated by \code{\link[stats]{cor}} or \code{\link[stats]{dist}} from the "stats" package respectively. The argument \code{p} sets the proportion of top most similar genes. Only genes within this proportion are selected.
}
\author{
Jianhai Zhang \email{zhang.jianhai@hotmail.com; jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
