% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_attr.R
\name{svg_attr}
\alias{svg_attr}
\title{Extract Attributes in an SVG File}
\usage{
svg_attr(doc, feature)
}
\arguments{
\item{doc}{An object of class "xml_document", i.e. the imported SVG file in R.}

\item{feature}{A character vector of features/samples extracted from the data. If some of the input features are duplicated in SVG file, then a reminder message is returned.}
}
\value{
A 3-component list of attribute data frame, the outline node, and the sample node.
}
\description{
This function extract basic attributes of nodes in an SVG file. The 'a' nodes are not deleted.
}
\references{
Hadley Wickham, Jim Hester and Jeroen Ooms (2019). xml2: Parse XML. R package version 1.2.2. https://CRAN.R-project.org/package=xml2
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
