% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.R
\name{video}
\alias{video}
\title{Make Spatial Heatmap Video}
\usage{
video(
  gg,
  cs.g,
  sam.uni,
  tis.trans,
  sub.title.size = NULL,
  bar.value.size = NULL,
  lgd.key.size = 0.02,
  lgd.text.size = 8,
  angle.text.key = NULL,
  position.text.key = NULL,
  lgd.row = 2,
  lgd.col = NULL,
  legend.value.vdo = NULL,
  label = FALSE,
  label.size = 4,
  label.angle = 0,
  hjust = 0,
  vjust = 0,
  opacity = 1,
  key = TRUE,
  width = 0.92,
  height = 0.99,
  video.dim = "640x480",
  res = 500,
  interval = 1,
  framerate = 1,
  out.dir
)
}
\arguments{
\item{gg}{A list of spatial heatmaps of ggplot.}

\item{cs.g}{The color key of ggplot.}

\item{sam.uni}{A vector of unique samples extracted from data matrix.}

\item{tis.trans}{A vector of tissues to be transparent.}

\item{sub.title.size}{A numeric. The subtitle font size of each individual spatial heatmap. Default is 11.}

\item{bar.value.size}{A numeric of value size in the color bar y-axis. Default is 10.}

\item{lgd.key.size}{The size of legend key (including text). Default is 0.02.}

\item{lgd.text.size}{The size of legend text. Default is 8.}

\item{angle.text.key}{A value of angle of key text in legend plot. Default is NULL, equivalent to 0.}

\item{position.text.key}{The position of key text in legend plot, one of "top", "right", "bottom", "left". Default is NULL, equivalent to "right".}

\item{lgd.row}{An integer of legend rows.}

\item{lgd.col}{An integer of legend columns.}

\item{legend.value.vdo}{Logical TRUE or FALSE. If TRUE, the numeric values of matching spatial features are added to legend, which applies to the video. The default is NULL.}

\item{label}{Logical. If TRUE, spatial features having matching samples are labeled by feature identifiers. Default is FALSE. It is useful when spatial features are labeled by similar colors.}

\item{label.size}{The size of spatial feature labels in legend plot. Default is 4.}

\item{label.angle}{The angle of spatial feature labels in legend plot. Default is 0.}

\item{hjust}{The value to horizontally adjust positions of spatial feature labels in legend plot. Default is 0.}

\item{vjust}{The value to vertically adjust positions of spatial feature labels in legend plot. Default is 0.}

\item{opacity}{The transparency of colored spatial features in legend plot. Default is 1. If 0, features are totally transparent.}

\item{key}{Logical. Default is TRUE and keys are added in legend plot.}

\item{width}{The image width in video in "npc", ranging from 0 to 0.92. Default is 0.92.}

\item{height}{The image height in video in "npc", ranging from 0 to 0.99. Default is 0.99.}

\item{video.dim}{A single character of the dimension of video frame in form of 'widthxheight', such as '1920x1080', '1280x800', '320x568', '1280x1024', '1280x720', '320x480', '480x360', '600x600', '800x600', '640x480'. Default is '640x480'.}

\item{res}{Resolution of the video in dpi.}

\item{interval}{The time interval (seconds) between spatial heatmap frames in the video. Default is 1.}

\item{framerate}{An integer of video framerate in frames per seconds. Default is 1. Larger values make the video smoother.}

\item{out.dir}{The directory to save interactive spatial heatmaps as independent HTML files and videos. Default is NULL, and the HTML files and videos are not saved.}
}
\value{
A video is saved in \code{out.dir}.
}
\description{
Make Spatial Heatmap Video
}
\references{
Jeroen Ooms (2020). av: Working with Audio and Video in R. R package version 0.5.0. https://CRAN.R-project.org/package=av
Baptiste Auguie (2017). gridExtra: Miscellaneous Functions for "Grid" Graphics. R package version 2.3. https://CRAN.R-project.org/package=gridExtra
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
