% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_mod.R
\name{adj_mod}
\alias{adj_mod}
\title{Compute Adjacency Matrix and Identify Modules}
\usage{
adj_mod(data, type, minSize = 15, dir = NULL)
}
\arguments{
\item{data}{The processed data matrix and metadata (optional) resulting from the function "filter_data", which is a "SummarizedExperiment" object.}

\item{type}{There are two options in this argument, "signed" and "unsigned". "signed" means both positive and negative adjacency between genes are maintained in network module identification while "unsigned" takes the absolute values of negative adjacency. Refer to "WGCNA" (Langfelder and Horvath 2008) for more details.}

\item{minSize}{The minimum module size in module identification. The default value is 15. Refer to "WGCNA" for more details.}

\item{dir}{The path where the directory "local_mode_result/" is created automatically to save resulting files "adj.txt" and "mod.txt". This argument should be the same with that from the function "filter_data" so that the processed data matrix generated by "filter_data" is saved in the same location. The default is NULL.}
}
\value{
A list containing the adjacency matrix and module assignment. If "dir" is not NULL, both are automatically saved in the directory "local_mode_result" as "adj.txt" and "mod.txt" respectively.
}
\description{
It is designed to compute adjacency matrix and identify modules for the filtered expression matrix. The computation is computationally intensive for large gene expression matrix (e.g. > 10,000 genes). Therefore, when using the web-browser based spatialHeatmap, the large expression matrix should be first computed locally using this function then the resulting files should be uploaded to "Compute locally" in the app.
}
\section{Details}{

This function builds on "WGCNA" (Langfelder and Horvath 2008) and flashClust (Langfelder and Horvath 2012). First, it computes the adjacency matrix based on the input data matrix, then converts the resulting adjacency matrix to a scale-free topology using a soft threshold (sft). If the type is "signed" sft=12, if the type is "unsigned" sft=6. Next, the modules are identified using dynamic tree cutting algorithm at four sensitivity levels: 0, 1, 2, 3. From 0 to 3, more modules are identified but module sizes are decreasing. In order to enhance the interactive performance of the network feature, modules are only identified at level 2 and 3, since the interactive network functionality in this package works better on smaller modules. Refer to "WGCNA" for more details.
}

\examples{
# Creat the "SummarizedExperiment" class. Refer to the R package "SummarizedExperiment" for more details.
data.path <- system.file("extdata/shinyApp/example", "root_expr_row_gen.txt", package = "spatialHeatmap")   
## The expression matrix.
library(data.table); expr <- fread(data.path, sep='\\t', header=TRUE, fill=TRUE)
col.na <- colnames(expr)[-ncol(expr)]; row.na <- as.data.frame(expr[, 1])[, 1]
expr <- as.matrix(as.data.frame(expr, stringsAsFactors=FALSE)[, -1])
rownames(expr) <- row.na; colnames(expr) <- col.na
## The expression matrix is stored in a "SummarizedExperiment" object.
library(SummarizedExperiment); expr <- SummarizedExperiment(assays=list(expr=expr))  
exp <- filter_data(data=expr, pOA=c(0, 0), CV=c(0.1, Inf), dir=NULL) 
adj_mod <- adj_mod(data=exp, type="signed", minSize=15, dir=NULL)
}
\references{
Langfelder P and Horvath S, WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics 2008, 9:559 doi:10.1186/1471-2105-9-559 \cr Peter Langfelder, Steve Horvath (2012). Fast R Functions for Robust Correlations and Hierarchical Clustering. Journal of Statistical Software, 46(11), 1-17. URL http://www.jstatsoft.org/v46/i11/ \cr R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/ \cr Peter Langfelder, Bin Zhang and with contributions from Steve Horvath (2016). dynamicTreeCut: Methods for Detection of Clusters in Hierarchical Clustering Dendrograms. R package version 1.63-1. https://CRAN.R-project.org/package=dynamicTreeCut \cr Martin Morgan, Valerie Obenchain, Jim Hester and Hervé Pagès (2018). SummarizedExperiment: SummarizedExperiment container. R package version 1.10.1
}
\author{
Jianhai Zhang \email{zhang.jianhai@hotmail.com; jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
