% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_mod.R
\name{adj_mod}
\alias{adj_mod}
\title{Compute Adjacency Matrix and Identify Modules}
\usage{
adj_mod(se, type, minSize = 15, dir = NULL)
}
\arguments{
\item{se}{A "SummarizedExperiment" object containing a data matrix and metadata returned by the function \code{\link{filter_data}}.}

\item{type}{"signed" or "unsigned". The "signed" means both positive and negative adjacency between genes are maintained in network module identification while "unsigned" takes the absolute values of negative adjacency. Refer to "WGCNA" (Langfelder and Horvath 2008) for more details.}

\item{minSize}{The minimum module size in module identification. The default value is 15. Refer to "WGCNA" for more details.}

\item{dir}{The directory path where the folder "local_mode_result/" is created automatically to save adjacency matrix and module assignment as tab-sepatated "adj.txt" and "mod.txt" respectively. This parameter should be the same with that from the function \code{\link{filter_data}} so that the "processed_data.txt" generated by \code{\link{filter_data}} is saved in the same location. The default is NULL. This parameter is designed since the computation is intensive for large data matrix (e.g. > 10,000 genes). Therefore, to avoid repetitive computation when using the Shiny app launched by \code{\link{shiny_all}}, "adj.txt" and "mod.txt" can be saved and uploaded to "Compute locally" at a later time.}
}
\value{
A list containing the adjacency matrix and module assignment. If "dir" is not NULL, both are automatically saved in the directory "local_mode_result" as "adj.txt" and "mod.txt" respectively, which can be uploaded to \code{\link{shiny_all}}.
}
\description{
It is designed to compute adjacency matrix and identify modules using the filtered expression data matrix from \code{\link{filter_data}}. The results should be provided to \code{\link{matrix_hm}} and \code{\link{network}}.
}
\section{Details}{

This function builds on "WGCNA" (Langfelder and Horvath 2008) and flashClust (Langfelder and Horvath 2012). First, it computes the adjacency matrix based on the input data matrix, then converts the resulting adjacency matrix to a scale-free topology using a soft threshold (sft). If the type is "signed" sft=12, if the type is "unsigned" sft=6. Next, the modules are identified using dynamic tree cutting algorithm at four sensitivity levels (ds): 0, 1, 2, 3. From 0 to 3, more modules are identified but module sizes are decreasing. In order to enhance the interactive performance of the network feature, modules are only identified at level 2 and 3, since the interactive network in this package works better on smaller modules. Refer to "WGCNA" for more details.
}

\examples{
# The example data (E-GEOD-67196) is an RNA-seq data measured in cerebellum and frontal cortex of human brain across normal and amyotrophic lateral sclerosis (ALS) subjects (Prudencio et al. 2015). 
library(ExpressionAtlas); library(SummarizedExperiment)
rse.hum <- getAtlasData('E-GEOD-67196')[[1]][[1]]; assay(rse.hum)[1:3, 1:3]

# A targets file describing replicates of samples and conditions is required, which is made based on the "colData" slot in the downloaded "RangedSummarizedExperiment" and available in spatialHeatmap. See the "se" parameter for details. 
brain.pa <- system.file('extdata/shinyApp/example/target_brain.txt', package='spatialHeatmap')
target.hum <- read.table(brain.pa, header=TRUE, row.names=1, sep='\t')
# The "organism_part" and "disease" column describes tissue and condition replicates respectively.  
target.hum[c(1:3, 41:42), 4:5]
# Place the targets file into "colData" slot as a DataFrame class. 
colData(rse.hum) <- DataFrame(target.hum)

# For users with little R expertise, if the gene expression matrix comes as a data frame, it should be placed into "SummarizedExperiment" before proceeding to next step. An example is shown below by borrowing a data matrix from the brain data.
# Borrow a data matrix.
df <- assay(rse.hum); df[1:2, 1:3]
# Place the data matrix and targets file (target.hum) into "SummarizedExperiment".
rse.hum <- SummarizedExperiment(assay=df, colData=target.hum, rowData=NULL)

# The count matrix is normalised with estimateSizeFactors (type="ratio").
se.nor.hum <- norm_data(se=rse.hum, method.norm='ratio', data.trans='log2')

# Average replicates of concatenated sample__condition.
se.aggr.hum <- aggr_rep(se=se.nor.hum, sam.factor='organism_part', con.factor='disease', aggr='mean')
assay(se.aggr.hum)[49939:49942, ] # The concatenated tissue__conditions are the column names of the output data matrix.

# Genes with low expression level and low variantion are always filtered. 
se.fil.hum <- filter_data(se=se.aggr.hum, sam.factor='organism_part', con.factor='disease', pOA=c(0.01, 5), CV=c(0.3, 100), dir=NULL)

# Detect modules. 
adj.mod <- adj_mod(se=se.fil.hum, type="signed", minSize=15, dir=NULL)
# The first column is ds=2 while the second is ds=3. The numbers in each column are module labels with "0" meaning genes not assigned to any modules.
adj.mod[['mod']][1:3, ]
}
\references{
Langfelder P and Horvath S, WGCNA: an R package for weighted correlation network analysis. BMC Bioinformatics 2008, 9:559 doi:10.1186/1471-2105-9-559 \cr Peter Langfelder, Steve Horvath (2012). Fast R Functions for Robust Correlations and Hierarchical Clustering. Journal of Statistical Software, 46(11), 1-17. URL http://www.jstatsoft.org/v46/i11/ \cr R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/ \cr Peter Langfelder, Bin Zhang and with contributions from Steve Horvath (2016). dynamicTreeCut: Methods for Detection of Clusters in Hierarchical Clustering Dendrograms. R package version 1.63-1. https://CRAN.R-project.org/package=dynamicTreeCut \cr Martin Morgan, Valerie Obenchain, Jim Hester and Hervé Pagès (2018). SummarizedExperiment: SummarizedExperiment container. R package version 1.10.1 
Prudencio, Mercedes, Veronique V Belzil, Ranjan Batra, Christian A Ross, Tania F Gendron, Luc J Pregent, Melissa E Murray, et al. 2015. "Distinct Brain Transcriptome Profiles in C9orf72-Associated and Sporadic ALS." Nat. Neurosci. 18 (8): 1175–82
Keays, Maria. 2019. ExpressionAtlas: Download Datasets from EMBL-EBI Expression Atlas
Love, Michael I., Wolfgang Huber, and Simon Anders. 2014. "Moderated Estimation of Fold Change and Dispersion for RNA-Seq Data with DESeq2." Genome Biology 15 (12): 550. doi:10.1186/s13059-014-0550-8
}
\author{
Jianhai Zhang \email{zhang.jianhai@hotmail.com; jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
