% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_asvg_data.R
\docType{data}
\name{aSVG.remote.repo}
\alias{aSVG.remote.repo}
\title{A list of URLs of remote aSVG repos}
\format{
A list.
}
\source{
\href{https://github.com/ebi-gene-expression-group/anatomogram/tree/master/src/svg}{EBI anatomogram}
\href{https://github.com/jianhaizhang/SHM_SVG/tree/master/SHM_SVG_repo}{spatialHeatmap_aSVG_Repository}
}
\usage{
data(aSVG.remote.repo)
}
\description{
A list of URLs of remote aSVG repos, \emph{i.e. EBI anatomogram and spatialHeatmap_aSVG_Repository}.
}
\examples{
data(aSVG.remote.repo)
aSVG.remote.repo
}
\references{
https://github.com/ebi-gene-expression-group/anatomogram/tree/master/src/svg
https://github.com/jianhaizhang/SHM_SVG/tree/master/SHM_SVG_repo
}
\keyword{datasets}
