% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_bar.R
\name{auc_bar}
\alias{auc_bar}
\title{Bar plot of AUCs from validating data sets}
\usage{
auc_bar(
  df.auc,
  auc = "auc",
  thr = 0.5,
  bar.width = 0.8,
  spd.sel = NULL,
  title = NULL,
  key.title = NULL,
  x.agl = 80,
  x.vjust = 0.6
)
}
\arguments{
\item{df.auc}{The \code{data.frame} of AUCs generated by optimal parameter settings on validating data sets, which is returned by \code{cocluster}.}

\item{auc}{The column name of AUCs in \code{df.auc}.}

\item{thr}{The AUC threshold, which will be labeled in the bar plot.}

\item{bar.width}{Width of a single bar.}

\item{spd.sel}{A character vector of selected \code{spd.set}, usually the common \code{spd.set} displaying desired AUCs across multiple validating data sets.}

\item{title}{The title of composite violin plots.}

\item{key.title}{The title of legend.}

\item{x.agl, x.vjust}{The angle and vertical position of x-axis text.}
}
\value{
An object of ggplot.
}
\description{
After coclustering optimization, visualize AUCs generated by optimal parameter settings on validating data sets.
}
\examples{
See package vignette by calling "browseVignettes('spatialHeatmap')".
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
