% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_bar.R
\name{auc_bar}
\alias{auc_bar}
\title{Bar plot of AUCs from validating data sets}
\usage{
auc_bar(
  df.auc,
  auc = "auc",
  thr = 0.5,
  bar.width = 0.8,
  spd.sel = NULL,
  title = NULL,
  title.size = 15,
  xlab = "Optimized settings",
  ylab = "AUC",
  axis.title.size = 11,
  x.text.size = 15,
  y.text.size = 15,
  x.agl = 80,
  x.vjust = 0.6
)
}
\arguments{
\item{df.auc}{The \code{data.frame} of AUCs generated by optimal parameter settings on validating data sets, which is returned by \code{cocluster}.}

\item{auc}{The column name of AUCs in \code{df.auc}.}

\item{thr}{The AUC threshold, which will be labeled in the bar plot.}

\item{bar.width}{Width of a single bar.}

\item{spd.sel}{A character vector of selected \code{spd.set}, usually the common \code{spd.set} displaying desired AUCs across multiple validating data sets.}

\item{title}{The title of composite violin plots.}

\item{title.size}{The title size. Default is 20.}

\item{axis.title.size}{The size of x and y axis labels.}

\item{x.text.size, y.text.size}{The size of x and y axis text.}

\item{x.agl, x.vjust}{The angle and vertical position of x-axis text.}

\item{x.lab, ylab}{The x and y axis label respectively.}
}
\value{
An object of ggplot.
}
\description{
After coclustering optimization, visualize AUCs generated by optimal parameter settings on validating data sets.
}
\examples{
# Create parameter settings and corresponding aucs.
df.auc <- data.frame(auc=runif(n=5, min=1e-12, max=.99))
ran <- seq(0.2, 0.8, 0.1)
df.spd <- data.frame(sim=sample(ran, 5, replace=FALSE), sim.p=sample(ran, 5, replace=FALSE), dim=sample(seq(5, 40, 1), 5, replace=FALSE))
df.spd$spd.set <- paste0('s', df.spd$sim, 'p', df.spd$sim.p, 'd', df.spd$dim)

df.auc <- cbind(df.spd, df.auc)
# Plot the aucs.
auc_bar(df.auc, auc='auc', thr=0.5) 
# See package vignette by calling "browseVignettes('spatialHeatmap')".
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
