% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_bar.R
\name{col_bar}
\alias{col_bar}
\title{Build Color Bar for the Spatial Heatmaps}
\usage{
col_bar(
  geneV,
  cols,
  width,
  bar.title.size = 0,
  bar.value.size = 10,
  mar = c(0.07, 0.01, 0.05, 0.1)
)
}
\arguments{
\item{geneV}{The gene expression values used for building the color bar.}

\item{cols}{The color codes corresponding to geneV.}

\item{width}{A numeric between 0 and 1, used to reduce the widths of bins in the color bar so as to avoid overlaps.}

\item{bar.title.size}{A numeric of color bar title size. The default is 0.}

\item{bar.value.size}{A numeric of value size in the color bar y-axis. The default is 10.}

\item{mar}{A 4-length numeric vector specifying margins of the color bar, in the "top", "right", "bottom", "left" order. The unit is "cm". Default is "c(3, 0.1, 3, 0.1)".}
}
\value{
An image of ggplot.
}
\description{
Build Color Bar for the Spatial Heatmaps
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
