% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com_factor.R
\name{com_factor}
\alias{com_factor}
\title{Combine Factors in Targets File}
\usage{
com_factor(se, target, factors2com, sep = ".", factor.new)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object.}

\item{target}{A \code{data.frame} object of targets file.}

\item{factors2com}{A character vector of column names or a numeric vector of column indeces in the targets file. Entries in these columns are combined.}

\item{sep}{The separator in the combined factors. One of \code{_}, and \code{.} (default).}

\item{factor.new}{The column name of the new combined factors.}
}
\value{
If \code{se} is provided, a \code{SummarizedExperiment} object is returned, where the \code{colData} slot contains the new column of combined factors. Otherwise, a\code{data.frame} object is returned, where the new column of combined factors is appended.
}
\description{
This is a helper function for data/aSVGs involving three or more factors such as sample, time, condition. It combine factors in targets file to make composite factors.
}
\examples{
clp.tar <- system.file('extdata/shinyApp/example/target_coleoptile.txt', package='spatialHeatmap')
target.clp <- read_fr(clp.tar)
target.clp <- com_factor(target=target.clp, factors2com=c('organism_part', 'age'), factor.new='samTime')
}
\references{
Narsai, Reena, David Secco, Matthew D Schultz, Joseph R Ecker, Ryan Lister, and James Whelan. 2017. "Dynamic and Rapid Changes in the Transcriptome and Epigenome During Germination and in Developing Rice (Oryza Sativa) Coleoptiles Under Anoxia and Re-Oxygenation." Plant J. 89 (4): 805–24
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
