% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter.data}
\alias{filter.data}
\title{Filter the Data Matrix}
\usage{
filter.data(data, sep, isRowGene, pOA = c(0, 0), CV = c(0, 10000),
  name)
}
\arguments{
\item{data}{The path of the data matrix. In the example of gene expression matrix, the dimension names are gene IDs and sample/conditions. The sample/condition names MUST be fomatted this way: a sample name is followed by double underscore then the condition, such as "sample__condition". The meta data (e.g. gene annotation) can also be included in parallel with sample/condition. In the names of sample/condition and meta data, only letters, digits, single underscore, dots are allowed. \cr E.g.: system.file("extdata/example", "root_expr_ann_row_gen.txt", package = "spatialHeatmap").}

\item{sep}{The seprator of the data matrix, e.g. ",", "\\t", ";".}

\item{isRowGene}{It specifies if the row names are used to display spatial heatmaps. The options are "TRUE" or "FALSE". For example, in a gene expression matrix genes are used to display heatmaps and the gene IDs are rows names, then the option is "TRUE".}

\item{pOA}{It specifies parameters of a filter function that filters according to the proportion of elements exceeding a threshold A. The input is a two-component vector, where the first one is the proportion and the second one is A, e.g.: c(0.1, 2). The default is c(0, 0), which means no filter is applied. Refer to "pOverA" from the package "genefilter".}

\item{CV}{It specifies parameters of a filter function that filters according to the coefficient of variation (CV). The input is a two-component vector, where the first and second mean the lower and upper bound of CVs used to filter, e.g.: cv(0.1, 5). The default is cv(0, 10000), which tries to aviod filtering.  Refer to "cv" from the package "genefilter".}

\item{name}{The processed data matrix is a "\\t" separated "txt" file, and this argument is the name of the "txt file". E.g if name="processed_expression", the resulting file is "processed_expression.txt".}
}
\value{
A data matrix filtered by "pOverA" and "cv" from the R package "genefilter", which is saved in the directory "local_mode_result". It is a "\\t" separated "txt" file and row names (e.g.: gene IDs) will be used to display the spatial heatmaps.
}
\description{
It designed to filter large data matrix (e.g.: > 10,000 rows) locally. The two main filter functions are "pOverA" and "cv", which are from the package "genefilter". The processed data matrix (processed_data.txt) is saved in the directory "local_mode_result", which should be uploaded to "Step 2: upload a gene expression file" in the Shiny app, which is launched by running "spatial.hm.all()".
}
\examples{
data.path <- system.file("extdata/example", "root_expr_ann_row_gen.txt", package = "spatialHeatmap")
exp <- filter.data(data=data.path, sep="\\t", isRowGene=TRUE, pOA=c(0, 0), CV=c(0.1, 10000), "processed_data")
}
\references{
R. Gentleman, V. Carey, W. Huber and F. Hahne (2017). genefilter: genefilter: methods for filtering genes from high-throughput experiments. R package version 1.58.1 \cr Matt Dowle and Arun Srinivasan (2017). data.table: Extension of `data.frame`. R package version 1.10.4. https://CRAN.R-project.org/package=data.table
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
