% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_lgd.R
\name{gg_lgd}
\alias{gg_lgd}
\title{Adjust Legend Key Size and Rows in Ggplot.}
\usage{
gg_lgd(
  gg.all,
  size.key = NULL,
  size.text.key = 8,
  angle.text.key = NULL,
  position.text.key = NULL,
  legend.value.vdo = NULL,
  sub.title.size = NULL,
  row = NULL,
  col = NULL,
  label = FALSE,
  label.size = 3,
  label.angle = 0,
  hjust = 0,
  vjust = 0,
  opacity = 1,
  key = TRUE,
  sam.dat,
  tis.trans = NULL
)
}
\arguments{
\item{gg.all}{A list of spatial heatmaps of ggplot.}

\item{size.key}{A numeric of legend key size. If \code{size.text} is NULL, it also applies to legend text size.}

\item{size.text.key}{A numeric of legend text size.}

\item{angle.text.key}{A value of angle of key text in legend plot. Default is NULL, equivalent to 0.}

\item{position.text.key}{The position of key text in legend plot, one of "top", "right", "bottom", "left". Default is NULL, equivalent to "right".}

\item{legend.value.vdo}{Logical TRUE or FALSE. If TRUE, the numeric values of matching spatial features are added to legend, which applies to the video. The default is NULL.}

\item{sub.title.size}{The title size of ggplot.}

\item{row}{An integer of rows in legend key.}

\item{col}{An integer of columns in legend key.}

\item{label}{Logical. If TRUE, spatial features having matching samples are labeled by feature identifiers. Default is FALSE. It is useful when spatial features are labeled by similar colors.}

\item{label.size}{The size of spatial feature labels in legend plot. Default is 4.}

\item{label.angle}{The angle of spatial feature labels in legend plot. Default is 0.}

\item{hjust}{The value to horizontally adjust positions of spatial feature labels in legend plot. Default is 0.}

\item{vjust}{The value to vertically adjust positions of spatial feature labels in legend plot. Default is 0.}

\item{opacity}{The transparency of colored spatial features in legend plot. Default is 1. If 0, features are totally transparent.}

\item{key}{Logical. Default is TRUE and keys are added in legend plot.}

\item{sam.dat}{A vector of samples in the data matrix.}

\item{tis.trans}{A vector of tissues to be transparent.}
}
\value{
A list of ggplots.
}
\description{
Adjust Legend Key Size and Rows in Ggplot.
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
