% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_list.R
\name{grob_list}
\alias{grob_list}
\title{Draw each spatial heatmap and convert them to ggplot2 plot grob}
\usage{
grob_list(
  gene,
  geneV,
  coord,
  ID,
  cols,
  tis.path,
  tis.trans = NULL,
  sub.title.size,
  sam.legend = "identical",
  legend.col = NULL,
  legend.title = NULL,
  legend.ncol = NULL,
  legend.nrow = NULL,
  legend.position = "bottom",
  legend.direction = NULL,
  legend.key.size = 0.5,
  legend.label.size = 8,
  legend.title.size = 8,
  line.size = 0.2,
  line.color = "grey70",
  ...
)
}
\arguments{
\item{gene}{The gene expession matrix, where rows are genes and columns are tissue/conditions.}

\item{geneV}{The gene expression values used to construct the colour bar.}

\item{coord}{The coordidates extracted from the SVG file.}

\item{ID}{All gene ids selected after the App is launched.}

\item{cols}{All the colour codes used to construct the colour bar.}

\item{tis.path}{All the tissues/paths extracted from the SVG.}

\item{tis.trans}{A character vector of tissue names. These tissues cover other tissues and should be set transparent. E.g c("epidermis", "cortex").}

\item{sub.title.size}{A numeric. The subtitle font size of each individual spatial heatmap. Default is 11.}

\item{sam.legend}{A character vector of tissue names in the SVG image. These tissues are shown in the legend. Default is "identical", meaning all the identical tissues between the data matrix and SVG image. Another value is 'all', meaning all tissues in the SVG image.}

\item{legend.col}{A character vector of colours provided by users for the legend keys. The lenght must be equal to the number of target samples shown in the legend. Default is NULL, which means colours are chosen automatically.}

\item{legend.title}{A character, the legend title. Default is NULL.}

\item{legend.ncol}{An integer, the column number of the items in the legend. Default is NULL.}

\item{legend.nrow}{An integer, the row number of the items in the legend. Default is NULL.}

\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}

\item{legend.direction}{layout of items in legends ("horizontal" or
"vertical")}

\item{legend.key.size}{A numeric (in "cm"). Default is 0.5. Size of the legend key.}

\item{legend.label.size}{A numeric. Default is 8. Size of the legend label.}

\item{legend.title.size}{A numeric. Default is 8. Size of the legend title.}

\item{line.size}{A numeric. The size of the polygon outline. Default is 0.2.}

\item{line.color}{A character. The color of polygon outline. Default is "grey70".}

\item{...}{Other arguments passed to ggplot().}
}
\value{
A list of spatial heatmaps in the form of ggplot2 plot grob.
}
\description{
Draw each spatial heatmap and convert them to ggplot2 plot grob
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016. \cr R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. RL https://www.R-project.org/ \cr Mustroph, Angelika, M Eugenia Zanetti, Charles J H Jang, Hans E Holtan, Peter P Repetti, David W Galbraith, Thomas Girke, and Julia Bailey-Serres. 2009. “Profiling Translatomes of Discrete Cell Populations Resolves Altered Cellular Priorities During Hypoxia in Arabidopsis.” Proc Natl Acad Sci U S A 106 (44): 18843–8
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
