% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_list.R
\name{grob_list}
\alias{grob_list}
\title{Draw each spatial heatmap and convert them to ggplot2 plot grob}
\usage{
grob_list(
  gene,
  geneV,
  coord,
  ID,
  cols,
  tis.path,
  tis.trans = NULL,
  sub.title.size,
  ...
)
}
\arguments{
\item{gene}{The gene expession matrix, where rows are genes and columns are tissue/conditions.}

\item{geneV}{The gene expression values used to construct the colour bar.}

\item{coord}{The coordidates extracted from the SVG file.}

\item{ID}{All gene ids selected after the App is launched.}

\item{cols}{All the colour codes used to construct the colour bar.}

\item{tis.path}{All the tissues/paths extracted from the SVG.}

\item{tis.trans}{The tissues selected by users that are expected to be transparent.}

\item{sub.title.size}{The subtitle font size of each individual spatial heatmap.}

\item{...}{Other arguments passed to ggplot().}
}
\value{
A list of spatial heatmaps in the form of ggplot2 plot grob.
}
\description{
Draw each spatial heatmap and convert them to ggplot2 plot grob
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016. \cr R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. RL https://www.R-project.org/ \cr Mustroph, Angelika, M Eugenia Zanetti, Charles J H Jang, Hans E Holtan, Peter P Repetti, David W Galbraith, Thomas Girke, and Julia Bailey-Serres. 2009. “Profiling Translatomes of Discrete Cell Populations Resolves Altered Cellular Priorities During Hypoxia in Arabidopsis.” Proc Natl Acad Sci U S A 106 (44): 18843–8
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
