% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob_list.R
\name{grob_list}
\alias{grob_list}
\title{Draw Each Spatial Heatmap and Convert Them to Ggplot2 Plot Grob}
\usage{
grob_list(
  gene,
  con.na = TRUE,
  geneV,
  coord,
  ID,
  cols,
  tis.path,
  tis.trans = NULL,
  sub.title.size,
  sam.legend = "identical",
  legend.col,
  legend.title = NULL,
  legend.ncol = NULL,
  legend.nrow = NULL,
  legend.position = "bottom",
  legend.direction = NULL,
  legend.key.size = 0.02,
  legend.label.size = 8,
  legend.title.size = 8,
  line.size = 0.2,
  line.color = "grey70",
  mar.lb = NULL,
  ...
)
}
\arguments{
\item{gene}{The gene expession matrix, where rows are genes and columns are tissue/conditions.}

\item{con.na}{Logical, TRUE or FALSE. Default is TRUE, meaning conditions are available.}

\item{geneV}{The gene expression values used to construct the colour bar.}

\item{coord}{The coordidates extracted from the SVG file.}

\item{ID}{All gene ids selected after the App is launched.}

\item{cols}{All the color codes used to construct the color bar.}

\item{tis.path}{All the tissues/paths extracted from the SVG.}

\item{tis.trans}{A character vector of tissue/spatial feature identifiers. These tissues may cover other tissues and should be set transparent. \emph{E.g} c("brain", "heart").}

\item{sub.title.size}{A numeric. The subtitle font size of each individual spatial heatmap. Default is 11.}

\item{sam.legend}{"identical", "all", or a character vector of tissue names from the aSVG image to show in the legend plot. Default is "identical", meaning all the identical/matching tissues between the data matrix and aSVG image. If "all", all tissues in the aSVG image are shown.}

\item{legend.col}{A character vector of colors for the legend keys. The lenght must be equal to the number of target samples shown in the legend.}

\item{legend.title}{A character, the legend title. Default is NULL.}

\item{legend.ncol}{An integer, the total columns of items in the legend. Default is NULL.}

\item{legend.nrow}{An integer, the total rows of the items in the legend. Default is NULL. It is applicable to both static image and video.}

\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}

\item{legend.direction}{layout of items in legends ("horizontal" or
"vertical")}

\item{legend.key.size}{A numeric (in "npc"). Default is 0.02. Size of the legend key, applicable to both static image and video.}

\item{legend.label.size}{A numeric. Default is 8. Size of the legend label, applicable to both static image and video.}

\item{legend.title.size}{A numeric. Default is 15. Size of the legend plot title.}

\item{line.size}{A numeric. The size of the shape outlines. Default is 0.2.}

\item{line.color}{A character. The color of shape outlines. Default is "grey70".}

\item{mar.lb}{A two-component numeric vector. The first and second numeric is left/right and bottom/top margin (npc) respectively.}

\item{...}{Other arguments passed to \code{\link[ggplot2]{ggplot}}.}
}
\value{
A nested list of spatial heatmaps of ggplot2 plot grob, spatial heatmaps of ggplot, and legend plot of ggplot.
}
\description{
Draw Each Spatial Heatmap and Convert Them to Ggplot2 Plot Grob
}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016. \cr R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. RL https://www.R-project.org/ \cr Mustroph, Angelika, M Eugenia Zanetti, Charles J H Jang, Hans E Holtan, Peter P Repetti, David W Galbraith, Thomas Girke, and Julia Bailey-Serres. 2009. “Profiling Translatomes of Discrete Cell Populations Resolves Altered Cellular Priorities During Hypoxia in Arabidopsis.” Proc Natl Acad Sci U S A 106 (44): 18843–8
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
