% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_ly.R
\name{html_ly}
\alias{html_ly}
\title{Build Interactive Spatial Heatmap in HTML}
\usage{
html_ly(
  gg,
  cs.g,
  tis.trans,
  sam.uni,
  anm.width,
  anm.height,
  selfcontained = FALSE,
  out.dir
)
}
\arguments{
\item{gg}{A list of spatial heatmaps of ggplot.}

\item{cs.g}{The color key of ggplot.}

\item{tis.trans}{A character vector of tissue/spatial feature identifiers that will be set transparent. \emph{E.g} c("brain", "heart"). This argument is used when target features are covered by  overlapping features and the latter should be transparent.}

\item{sam.uni}{A vector of unique samples extracted from data matrix.}

\item{anm.width}{The width of spatial heatmaps in HTML files. Default is 650.}

\item{anm.height}{The height of spatial heatmaps in HTML files. Default is 550.}

\item{selfcontained}{Whether to save the HTML as a single self-contained file
(with external resources base64 encoded) or a file with external resources
placed in an adjacent directory.}

\item{out.dir}{The directory to save interactive spatial heatmaps as independent HTML files and videos. Default is NULL, and the HTML files and videos are not saved.}
}
\value{
HTML files of spatial heatmaps are saved in 'animaiton_shm'.
}
\description{
Build Interactive Spatial Heatmap in HTML
}
\references{
C. Sievert. Interactive Web-Based Data Visualization with R, plotly, and shiny. Chapman and Hall/CRC Florida, 2020.
Ramnath Vaidyanathan, Yihui Xie, JJ Allaire, Joe Cheng and Kenton Russell (2019). htmlwidgets: HTML Widgets for R. R package version 1.5.1. https://CRAN.R-project.org/package=htmlwidgets
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
