% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Network}
\usage{
network(geneID, data, adj.mod, ds = "3", adj.min = 0, con.min = 0,
  node.col = c("mediumorchid1", "chocolate4"), edge.col = c("yellow",
  "blue"), vertex.label.cex = 1, vertex.cex = 3, edge.cex = 10,
  layout = "circle", main = NULL, static = TRUE, ...)
}
\arguments{
\item{geneID}{A gene ID from the expression matrix.}

\item{data}{The processed data matrix and metadata (optional) resulting from the function "filter.data", which is a "SummarizedExperiment" object.}

\item{adj.mod}{A list of "adjacency matrix" and "modules" definitions returned by the function "adj.mod".}

\item{ds}{The module identification sensitivity, either 2 or 3. See function "adj.mod" for details. Used for static network.}

\item{adj.min}{Minimum adjacency between genes, edges with adjacency below which will be removed. Used for static network.}

\item{con.min}{Minimun connectivity of a gene, genes with connectivity below which will be removed. Used for static network.}

\item{node.col}{A vector of colour ingredients for constructing node colour scale in the static image. The default is c("mediumorchid1", "chocolate4").}

\item{edge.col}{A vector of colour ingredients for constructing edge colour scale in the static image. The default is c("yellow", "blue").}

\item{vertex.label.cex}{The size of node label. The default is 1.}

\item{vertex.cex}{The size of node in the static image. The default is 3.}

\item{edge.cex}{The size of edge in the static image. The default is 10.}

\item{layout}{The layout of the network in static image, either "circle" or "fr". The "fr" stands for force-directed layout algorithm by Fruchterman and Reingold. The default is "circle".}

\item{main}{The title in the static image.}

\item{static}{Either "TRUE" or "FALSE", which displays static or interactive network respectively.}

\item{...}{Other arguments passed to the generic function "plot", e.g: "asp=1".}
}
\value{
A static or interactive network display.
}
\description{
This function represents the input gene in the context of corresponding gene network module, where nodes are genes and edges are adjacencies between genes. The network can be dispayed in static or interactive mode. \cr The gene modules are identified at two alternative sensitivity levels (3, 2). See function "adj_mod" for details. The thicker edge denotes higher adjacency (coexpression similarity) between genes while larger node indicates higher gene connectivity (sum of a gene's adjacency with all its direct neighbours). \cr In the interactive mode, there is an interactive colour bar to denote gene connectivity. The colour ingredients must only be separated by comma, e.g. "yellow,purple,blue", which means gene connectivity increases from yellow to blue. If too many edges (e.g.: > 300) are displayed, the network could get stuck. So the "Input an adjacency threshold to display the adjacency network." option sets a threthold to filter out weak edges and all remaining edges are displayed. If not too many (e.g.: < 300), users can check "Yes" under "Display or not?", then the network will be displayed and would be responsive smoothly. To maintain acceptable performance, users are advised to choose a stringent threshold (e.g. 0.9) initially, then decrease the value gradually. The interactive feature allows users to zoom in and out, or drag a gene around. All the gene IDs in the network module are listed in "Select by id" in decreasing order according to gene connectivity. The selected gene ID is appended "_selected" as a label. By clicking an ID in this list, users can identify the corresponding gene in the network. If the input has gene annotations, then the annotation can be seen by hovering the cursor over a node. \cr The same module can also be displayed in the form of a matrix heatmap with the function "matrix.heatmap".
}
\examples{
# Creat the "SummarizedExperiment" class. Refer to the R package "SummarizedExperiment" for more details.
data.path <- system.file("extdata/shinyApp/example", "root_expr_row_gen.txt", package="spatialHeatmap")   
## The expression matrix, where the row and column names should be gene IDs and sample/conditions respectively.
library(data.table); expr <- fread(data.path, sep='\\t', header=TRUE, fill=TRUE)
col.na <- colnames(expr)[-ncol(expr)]; row.na <- as.data.frame(expr[, 1])[, 1]
expr <- as.matrix(as.data.frame(expr, stringsAsFactors=FALSE)[, -1])
rownames(expr) <- row.na; colnames(expr) <- col.na
library(SummarizedExperiment); expr <- SummarizedExperiment(assays=list(expr=expr)) # Metadata is not necessary.  
exp <- filter_data(data=expr, pOA=c(0, 0), CV=c(0.1, 100), dir=NULL) 
adj.mod <- adj_mod(data=exp, type="signed", minSize=15, dir=NULL)
# The gene "PSAC" is represented in the context of its network module in the form of an interactive network.
network(geneID="PSAC", data=exp, adj.mod=adj.mod, static=TRUE)
}
\references{
Martin Morgan, Valerie Obenchain, Jim Hester and Hervé Pagès (2018). SummarizedExperiment: SummarizedExperiment container. R package version 1.10.1 \cr Csardi G, Nepusz T: The igraph software package for complex network research, InterJournal, Complex Systems 1695. 2006. http://igraph.org \cr R Core Team (2018). R: A language and environment for statistical computing. R Foundation for Statistical Computing, Vienna, Austria. URL https://www.R-project.org/ \cr Winston Chang, Joe Cheng, JJ Allaire, Yihui Xie and Jonathan McPherson (2018). shiny: Web Application Framework for R. R package version 1.1.0. https://CRAN.R-project.org/package=shiny \cr Winston Chang and Barbara Borges Ribeiro (2018). shinydashboard: Create Dashboards with 'Shiny'. R package version 0.7.1. https://CRAN.R-project.org/package=shinydashboard \cr Almende B.V., Benoit Thieurmel and Titouan Robert (2018). visNetwork: Network Visualization using 'vis.js' Library. R package version 2.0.4. https://CRAN.R-project.org/package=visNetwork
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
