% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dim.R
\name{plot_dim}
\alias{plot_dim}
\title{Plotting single cells in reduced dimensionalities}
\usage{
plot_dim(sce, dim, color.by, row.sel = NULL, x.break = NULL, y.break = NULL)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object with reduced dimensions seen by \code{reducedDimNames(sce)}.}

\item{dim}{One of \code{PCA}, \code{UMAP}, \code{TSNE}, the method for reducing dimensionality.}

\item{row.sel}{A numeric vector of row numbers in the \code{colData} slot of \code{sce}. The cells corresponding to these rows are highlighted and plotted on top of other cells.}

\item{x.break, y.break}{Two numeric vectors for x, y axis breaks respectively. E.g. \code{seq(-10, 10, 2)}. The default is \code{NULL}.}

\item{coloy.by}{One of the column names in the \code{colData} slot of \code{sce}.}
}
\value{
An object of ggplot.
}
\description{
Plotting single cells in reduced dimensionalities
}
\examples{
See function "cocluster" by running "?cocluster".
}
\references{
Amezquita R, Lun A, Becht E, Carey V, Carpp L, Geistlinger L, Marini F, Rue-Albrecht K, Risso D, Soneson C, Waldron L, Pages H, Smith M, Huber W, Morgan M, Gottardo R, Hicks S (2020). “Orchestrating single-cell analysis with Bioconductor.” Nature Methods, 17, 137–145. https://www.nature.com/articles/s41592-019-0654-x
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
Morgan M, Obenchain V, Hester J, Pagès H (2021). SummarizedExperiment: SummarizedExperiment container. R package version 1.24.0, https://bioconductor.org/packages/SummarizedExperiment.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
