% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_feature.R
\name{return_feature}
\alias{return_feature}
\title{Return SVG Images Containing Provided Features and Species}
\usage{
return_feature(
  feature,
  species,
  keywords.all = TRUE,
  remote = FALSE,
  dir = NULL,
  desc = FALSE,
  match.only = TRUE,
  return.all = FALSE
)
}
\arguments{
\item{feature}{A vector of keywords (case insentitive) of target feature(s), which is used to select SVG images from a collection. E.g. c("frontal", "cortex").}

\item{species}{A vector of keywords (case insentitive) of a target species, which is used to select SVG images from a collection. E.g. c("homo sapiens").}

\item{keywords.all}{Logical, TRUE or FALSE. Default is TRUE. If TRUE, every returned hit will contain all the provided keywords in "feature" and "species". Otherwise, every returned hit contain at least 1 keyword in "feature" and at least 1 keyword in "species".}

\item{remote}{Logical, FALSE or TRUE. Default is FALSE. If TRUE, the remote SVG repository "https://github.com/jianhaizhang/SVG_tutorial_file/tree/master/svg_repo" is used for query.}

\item{dir}{The directory where the SVG images are available. If "remote" is TRUE, the returned SVG images are saved in this directory. Note, in this case existing SVG images with identical names as returned ones are overwritten. If "remote" is FALSE, user-provided SVG images should be saved in this directory for query.}

\item{desc}{Logical, FALSE or TRUE. Default is FALSE. If TRUE, the feature descriptions from the package "rols" (Laurent Gatto 2019) are added. If too many features are returned, this process takes a long time.}

\item{match.only}{Logical, TRUE or FALSE. If TRUE (default), only features matching the feature keywords are returned. If FALSE, all features in the matching species are returned, and the matching features are listed on the top of the data frame.}

\item{return.all}{Logical, FALSE or TRUE. Default is FALSE. If TRUE, all features together with all SVG images are returned, either remote or user-provided.}
}
\value{
A feature data frame with columns corresponding to feature id, feature ontology id, feature is index in an SVG image, and/or feature description.
}
\description{
Successful spatial heatmap plotting requires the feature identifiers of interest are identical between the data matrix and SVG image. This function parses a collection of SVG images and returns existing features related to the provided keywords. If users want to use their custom feature identifiers, \code{\link{update_feature}} should be used. Otherwise the returned featrue identifiers should be used to replace the counterparts in the data matrix. Features denote tissues, cells, etc.
}
\examples{
feature.df <- return_feature(feature='frontal cortex', species='homo sapiens', keywords.all=TRUE, desc=FALSE, return.all=FALSE, dir='.', remote=TRUE)
}
\references{
Laurent Gatto (2019). rols: An R interface to the Ontology Lookup Service. R package version 2.14.0. http://lgatto.github.com/rols/
Hadley Wickham, Jim Hester and Jeroen Ooms (2019). xml2: Parse XML. R package version 1.2.2. https://CRAN.R-project.org/package=xml2
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
