% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_bulk.R
\name{true_bulk}
\alias{true_bulk}
\title{Assign true bulk to cells in \code{colData} slot.}
\usage{
true_bulk(sce, df.match)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} of clustered single cell data.}

\item{df.match}{The matching table between cells and true bulk.}
}
\value{
A \code{SingleCellExperiment} object.
}
\description{
In co-clustering, assign true bulk to cells in \code{colData} slot.
}
\examples{
See the example in the "cocluster" function by running "?cocluster".
}
\references{
Morgan M, Obenchain V, Hester J, Pagès H (2021). SummarizedExperiment: SummarizedExperiment container. R package version 1.24.  0, https://bioconductor.org/packages/SummarizedExperiment.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
