% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_feature.R
\name{update_feature}
\alias{update_feature}
\title{Update SVG Images with Provided Features}
\usage{
update_feature(feature, dir)
}
\arguments{
\item{feature}{The data frame returned by \code{\link{return_feature}} with the user-provided features being the first column.}

\item{dir}{The directory where the SVG images are available. It should be same with "dir" in \code{\link{return_feature}}.}
}
\value{
Nothing is returned. The SVG images of interest in "dir" are updated with new features, and are ready for plotting spatial heatmaps.
}
\description{
Successful spatial heatmap plotting requires the feature identifiers of interest are identical between the data matrix and SVG image. This function is designed to replace existing features in SVG images with user-provided features. Note this function treats the first column in the feature data frame as user-provided features, so custom features must be the first column.
}
\examples{
feature.df <- return_feature(feature='frontal cortex', species='homo sapiens', keywords.all=TRUE, desc=FALSE, return.all=FALSE, dir='.', remote=TRUE)
ft <- c('frontal.cortex', 'prefrontal.cortex', 'prefrontal.cortex', 'frontal.cortex', 'frontal.cortex', 'prefrontal.cortex')
# Add new features to the first column.
feature.df.new <- cbind(featureNew=ft, feature.df) 
update_feature(feature=ft2, dir='.')
}
\references{
Hadley Wickham, Jim Hester and Jeroen Ooms (2019). xml2: Parse XML. R package version 1.2.2. https://CRAN.R-project.org/package=xml2
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
