% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.R
\name{video}
\alias{video}
\title{Make Spatial Heatmap Video}
\usage{
video(
  gg,
  cs.g,
  sam.uni,
  tis.trans,
  lgd.key.size = 0.02,
  lgd.text.size = 8,
  lgd.row = 2,
  width = 1,
  height = 1,
  video.dim = "640x480",
  interval = 1,
  out.dir
)
}
\arguments{
\item{gg}{A list of spatial heatmaps of ggplot.}

\item{cs.g}{The color key of ggplot.}

\item{sam.uni}{A vector of unique samples extracted from data matrix.}

\item{tis.trans}{A vector of tissues to be transparent.}

\item{lgd.key.size}{The size of legend key (including text). Default is 0.02.}

\item{lgd.text.size}{The size of legend text. Default is 8.}

\item{lgd.row}{An integer of legend rows.}

\item{width}{A numeric of overall width of all subplots, ranging between 0 and 1. The default is 1. It is applicable to spatial heatmaps in both static image and video.}

\item{height}{A numeric of overall height of all subplots, ranging between 0 and 1. The default is 1. It is applicable to spatial heatmaps in both static image and video.}

\item{video.dim}{A single character of the dimension of video frame in form of 'widthxheight', such as '1920x1080', '1280x800', '320x568', '1280x1024', '1280x720', '320x480', '480x360', '600x600', '800x600', '640x480'. Default is '640x480'.}

\item{interval}{The time interval (seconds) between spatial heatmap frames in the video. Default is 1.}

\item{out.dir}{The directory to save interactive spatial heatmaps as independent HTML files and videos. Default is NULL, and the HTML files and videos are not saved.}
}
\value{
A video is saved in \code{out.dir}.
}
\description{
Make Spatial Heatmap Video
}
\references{
Yihui Xie (2013). animation: An R Package for Creating Animations and Demonstrating Statistical Methods. Journal of Statistical Software, 53(1), 1-27. URL http://www.jstatsoft.org/v53/i01/.
}
\author{
Jianhai Zhang \email{jzhan067@ucr.edu; zhang.jianhai@hotmail.com} \cr Dr. Thomas Girke \email{thomas.girke@ucr.edu}
}
\keyword{Internal}
